Network.AWS.Lambda.AddEventSource

Description

Identifies an Amazon Kinesis stream as the event source for an AWS Lambda function. AWS Lambda invokes the specified function when records are posted to the stream. This is the pull model, where AWS Lambda invokes the function. For more information, go to AWS LambdaL How it Works in the AWS Lambda Developer Guide. This association between an Amazon Kinesis stream and an AWS Lambda function is called the event source mapping. You provide the configuration information (for example, which stream to read from and which AWS Lambda function to invoke) for the event source mapping in the request body. This operation requires permission for the iam:PassRole action for the IAM role. It also requires permission for the lambda:AddEventSource action.

http://docs.aws.amazon.com/lambda/latest/dg/API_AddEventSource.html

Request

data AddEventSource

Request constructor

addEventSource

Arguments

:: Text
aesEventSource

-> Text
aesFunctionName

-> Text
aesRole

-> AddEventSource

AddEventSource constructor.

The fields accessible through corresponding lenses are:

aesBatchSize :: Maybe Int

aesEventSource :: Text

aesFunctionName :: Text

aesParameters :: HashMap Text Text

aesRole :: Text

Request lenses

aesBatchSize :: Lens' AddEventSource (Maybe Int)

The largest number of records that AWS Lambda will give to your function in a single event. The default is 100 records.

aesEventSource :: Lens' AddEventSource Text

The Amazon Resource Name (ARN) of the Amazon Kinesis stream that is the event source. Any record added to this stream causes AWS Lambda to invoke your Lambda function. AWS Lambda POSTs the Amazon Kinesis event, containing records, to your Lambda function as JSON.

aesFunctionName :: Lens' AddEventSource Text

The Lambda function to invoke when AWS Lambda detects an event on the stream.

aesParameters :: Lens' AddEventSource (HashMap Text Text)

A map (key-value pairs) defining the configuration for AWS Lambda to use when reading the event source. Currently, AWS Lambda supports only the InitialPositionInStream key. The valid values are: TRIM_HORIZON and LATEST. The default value is TRIM_HORIZON. For more information, go to ShardIteratorType in the Amazon Kinesis Service API Reference.

aesRole :: Lens' AddEventSource Text

The ARN of the IAM role (invocation role) that AWS Lambda can assume to read from the stream and invoke the function.

Response

data AddEventSourceResponse

Response constructor

addEventSourceResponse :: AddEventSourceResponse

AddEventSourceResponse constructor.

The fields accessible through corresponding lenses are:

aesrBatchSize :: Maybe Int

aesrEventSource :: Maybe Text

aesrFunctionName :: Maybe Text

aesrIsActive :: Maybe Bool

aesrLastModified :: Maybe UTCTime

aesrParameters :: HashMap Text Text

aesrRole :: Maybe Text

aesrStatus :: Maybe Text

aesrUUID :: Maybe Text

Response lenses

aesrBatchSize :: Lens' AddEventSourceResponse (Maybe Int)

The largest number of records that AWS Lambda will POST in the invocation request to your function.

aesrEventSource :: Lens' AddEventSourceResponse (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Kinesis stream that is the source of events.

aesrFunctionName :: Lens' AddEventSourceResponse (Maybe Text)

The Lambda function to invoke when AWS Lambda detects an event on the stream.

aesrIsActive :: Lens' AddEventSourceResponse (Maybe Bool)

Indicates whether the event source mapping is currently honored. Events are only processes if IsActive is true.

aesrLastModified :: Lens' AddEventSourceResponse (Maybe UTCTime)

The UTC time string indicating the last time the event mapping was updated.

aesrParameters :: Lens' AddEventSourceResponse (HashMap Text Text)

The map (key-value pairs) defining the configuration for AWS Lambda to use when reading the event source.

aesrRole :: Lens' AddEventSourceResponse (Maybe Text)

The ARN of the IAM role (invocation role) that AWS Lambda can assume to read from the stream and invoke the function.

aesrStatus :: Lens' AddEventSourceResponse (Maybe Text)

The description of the health of the event source mapping. Valid values are: PENDING, OK, and PROBLEM:message. Initially this staus is PENDING. When AWS Lambda begins processing events, it changes the status to OK.

aesrUUID :: Lens' AddEventSourceResponse (Maybe Text)

The AWS Lambda assigned opaque identifier for the mapping.