Network.AWS.Lambda.CreateFunction

Description

Creates a new Lambda function. The function metadata is created from the request parameters, and the code for the function is provided by a .zip file in the request body. If the function name already exists, the operation will fail. Note that the function name is case-sensitive.

This operation requires permission for the 'lambda:CreateFunction' action.

http://docs.aws.amazon.com/lambda/latest/dg/API_CreateFunction.html

Request

data CreateFunction

Request constructor

createFunction

Arguments

:: Text
cfFunctionName

-> Runtime
cfRuntime

-> Text
cfRole

-> Text
cfHandler

-> FunctionCode
cfCode

-> CreateFunction

CreateFunction constructor.

The fields accessible through corresponding lenses are:

cfCode :: FunctionCode

cfDescription :: Maybe Text

cfFunctionName :: Text

cfHandler :: Text

cfMemorySize :: Maybe Natural

cfRole :: Text

cfRuntime :: Runtime

cfTimeout :: Maybe Natural

Request lenses

cfCode :: Lens' CreateFunction FunctionCode

A structure that includes ZipFile.

cfDescription :: Lens' CreateFunction (Maybe Text)

A short, user-defined function description. Lambda does not use this value. Assign a meaningful description as you see fit.

cfFunctionName :: Lens' CreateFunction Text

The name you want to assign to the function you are uploading. You can specify an unqualified function name (for example, Thumbnail) or you can specify Amazon Resource Name (ARN) of the function (for example, "arn:aws:lambda:us-west-2:account-id:function:ThumbNail"). AWS Lambda also allows you to specify only the account ID qualifier (for example, "account-id:Thumbnail"). Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 character in length. The function names appear in the console and are returned in the ListFunctions API. Function names are used to specify functions to other AWS Lambda APIs, such as Invoke.

cfHandler :: Lens' CreateFunction Text

The function within your code that Lambda calls to begin execution. For Node.js, it is the module-name.export value in your function.

cfMemorySize :: Lens' CreateFunction (Maybe Natural)

The amount of memory, in MB, your Lambda function is given. Lambda uses this memory size to infer the amount of CPU and memory allocated to your function. Your function use-case determines your CPU and memory requirements. For example, a database operation might need less memory compared to an image processing function. The default value is 128 MB. The value must be a multiple of 64 MB.

cfRole :: Lens' CreateFunction Text

The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources. For more information, see AWS Lambda: How it Works

cfRuntime :: Lens' CreateFunction Runtime

The runtime environment for the Lambda function you are uploading. Currently, Lambda supports only "nodejs" as the runtime.

cfTimeout :: Lens' CreateFunction (Maybe Natural)

The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

Response

data CreateFunctionResponse

Response constructor

createFunctionResponse :: CreateFunctionResponse

CreateFunctionResponse constructor.

The fields accessible through corresponding lenses are:

cfrCodeSize :: Maybe Integer

cfrDescription :: Maybe Text

cfrFunctionArn :: Maybe Text

cfrFunctionName :: Maybe Text

cfrHandler :: Maybe Text

cfrLastModified :: Maybe Text

cfrMemorySize :: Maybe Natural

cfrRole :: Maybe Text

cfrRuntime :: Maybe Runtime

cfrTimeout :: Maybe Natural

Response lenses

cfrCodeSize :: Lens' CreateFunctionResponse (Maybe Integer)

The size, in bytes, of the function .zip file you uploaded.

cfrDescription :: Lens' CreateFunctionResponse (Maybe Text)

The user-provided description.

cfrFunctionArn :: Lens' CreateFunctionResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the function.

cfrFunctionName :: Lens' CreateFunctionResponse (Maybe Text)

The name of the function.

cfrHandler :: Lens' CreateFunctionResponse (Maybe Text)

The function Lambda calls to begin executing your function.

cfrLastModified :: Lens' CreateFunctionResponse (Maybe Text)

The timestamp of the last time you updated the function.

cfrMemorySize :: Lens' CreateFunctionResponse (Maybe Natural)

The memory size, in MB, you configured for the function. Must be a multiple of 64 MB.

cfrRole :: Lens' CreateFunctionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources.

cfrRuntime :: Lens' CreateFunctionResponse (Maybe Runtime)

The runtime environment for the Lambda function.

cfrTimeout :: Lens' CreateFunctionResponse (Maybe Natural)

The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.