Network.AWS.Lambda.DeleteEventSourceMapping

Description

Removes an event source mapping. This means AWS Lambda will no longer invoke the function for events in the associated source.

This operation requires permission for the 'lambda:DeleteEventSourceMapping' action.

http://docs.aws.amazon.com/lambda/latest/dg/API_DeleteEventSourceMapping.html

Request

data DeleteEventSourceMapping

Request constructor

deleteEventSourceMapping

Arguments

:: Text
desmUUID

-> DeleteEventSourceMapping

DeleteEventSourceMapping constructor.

The fields accessible through corresponding lenses are:

desmUUID :: Text

Request lenses

desmUUID :: Lens' DeleteEventSourceMapping Text

The event source mapping ID.

Response

data DeleteEventSourceMappingResponse

Response constructor

deleteEventSourceMappingResponse :: DeleteEventSourceMappingResponse

DeleteEventSourceMappingResponse constructor.

The fields accessible through corresponding lenses are:

desmrBatchSize :: Maybe Natural

desmrEventSourceArn :: Maybe Text

desmrFunctionArn :: Maybe Text

desmrLastModified :: Maybe UTCTime

desmrLastProcessingResult :: Maybe Text

desmrState :: Maybe Text

desmrStateTransitionReason :: Maybe Text

desmrUUID :: Maybe Text

Response lenses

desmrBatchSize :: Lens' DeleteEventSourceMappingResponse (Maybe Natural)

The largest number of records that AWS Lambda will retrieve from your event source at the time of invoking your function. Your function receives an event with all the retrieved records.

desmrEventSourceArn :: Lens' DeleteEventSourceMappingResponse (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Kinesis stream that is the source of events.

desmrFunctionArn :: Lens' DeleteEventSourceMappingResponse (Maybe Text)

The Lambda function to invoke when AWS Lambda detects an event on the stream.

desmrLastModified :: Lens' DeleteEventSourceMappingResponse (Maybe UTCTime)

The UTC time string indicating the last time the event mapping was updated.

desmrLastProcessingResult :: Lens' DeleteEventSourceMappingResponse (Maybe Text)

The result of the last AWS Lambda invocation of your Lambda function.

desmrState :: Lens' DeleteEventSourceMappingResponse (Maybe Text)

The state of the event source mapping. It can be Creating, Enabled, Disabled, Enabling, Disabling, Updating, or Deleting.

desmrStateTransitionReason :: Lens' DeleteEventSourceMappingResponse (Maybe Text)

The reason the event source mapping is in its current state. It is either user-requested or an AWS Lambda-initiated state transition.

desmrUUID :: Lens' DeleteEventSourceMappingResponse (Maybe Text)

The AWS Lambda assigned opaque identifier for the mapping.