Network.AWS.Lambda.DeleteFunction

Description

Deletes the specified Lambda function code and configuration.

When you delete a function the associated access policy is also deleted. You will need to delete the event source mappings explicitly.

This operation requires permission for the 'lambda:DeleteFunction' action.

http://docs.aws.amazon.com/lambda/latest/dg/API_DeleteFunction.html

Request

data DeleteFunction

Request constructor

deleteFunction

Arguments

:: Text
dfFunctionName

-> DeleteFunction

DeleteFunction constructor.

The fields accessible through corresponding lenses are:

dfFunctionName :: Text

Request lenses

dfFunctionName :: Lens' DeleteFunction Text

The Lambda function to delete.

You can specify an unqualified function name (for example, Thumbnail) or you can specify Amazon Resource Name (ARN) of the function (for example, "arn:aws:lambda:us-west-2:account-id:function:ThumbNail"). AWS Lambda also allows you to specify only the account ID qualifier (for example, "account-id:Thumbnail"). Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 character in length.

Response

data DeleteFunctionResponse

Response constructor

deleteFunctionResponse :: DeleteFunctionResponse

DeleteFunctionResponse constructor.