Network.AWS.Lambda.GetAlias

Description

Returns the specified alias information such as the alias ARN, description, and function version it is pointing to. For more information, see Introduction to AWS Lambda Aliases .

This requires permission for the lambda:GetAlias action.

Creating a Request

getAlias

Arguments

:: Text
gaFunctionName

-> Text
gaName

-> GetAlias

Creates a value of GetAlias with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gaFunctionName - Function name for which the alias is created. An alias is a subresource that exists only in the context of an existing Lambda function so you must specify the function name. Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 characters in length.

gaName - Name of the alias for which you want to retrieve information.

data GetAlias

See: getAlias smart constructor.

Request Lenses

gaFunctionName :: Lens' GetAlias Text

Function name for which the alias is created. An alias is a subresource that exists only in the context of an existing Lambda function so you must specify the function name. Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 characters in length.

gaName :: Lens' GetAlias Text

Name of the alias for which you want to retrieve information.

Destructuring the Response

aliasConfiguration :: AliasConfiguration

Creates a value of AliasConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

acName - Alias name.

acFunctionVersion - Function version to which the alias points.

acAliasARN - Lambda function ARN that is qualified using the alias name as the suffix. For example, if you create an alias called BETA that points to a helloworld function version, the ARN is arn:aws:lambda:aws-regions:acct-id:function:helloworld:BETA .

acDescription - Alias description.

data AliasConfiguration

Provides configuration information about a Lambda function version alias.

See: aliasConfiguration smart constructor.

Response Lenses

acName :: Lens' AliasConfiguration (Maybe Text)

Alias name.

acFunctionVersion :: Lens' AliasConfiguration (Maybe Text)

Function version to which the alias points.

acAliasARN :: Lens' AliasConfiguration (Maybe Text)

Lambda function ARN that is qualified using the alias name as the suffix. For example, if you create an alias called BETA that points to a helloworld function version, the ARN is arn:aws:lambda:aws-regions:acct-id:function:helloworld:BETA .

acDescription :: Lens' AliasConfiguration (Maybe Text)

Alias description.