Network.AWS.Lambda.GetEventSource

Description

Returns configuration information for the specified event source mapping (see AddEventSource).

This operation requires permission for the 'lambda:GetEventSource' action.

http://docs.aws.amazon.com/lambda/latest/dg/API_GetEventSource.html

Request

data GetEventSource

Request constructor

getEventSource

Arguments

:: Text
gesUUID

-> GetEventSource

GetEventSource constructor.

The fields accessible through corresponding lenses are:

gesUUID :: Text

Request lenses

gesUUID :: Lens' GetEventSource Text

The AWS Lambda assigned ID of the event source mapping.

Response

data GetEventSourceResponse

Response constructor

getEventSourceResponse :: GetEventSourceResponse

GetEventSourceResponse constructor.

The fields accessible through corresponding lenses are:

gesrBatchSize :: Maybe Int

gesrEventSource :: Maybe Text

gesrFunctionName :: Maybe Text

gesrIsActive :: Maybe Bool

gesrLastModified :: Maybe UTCTime

gesrParameters :: HashMap Text Text

gesrRole :: Maybe Text

gesrStatus :: Maybe Text

gesrUUID :: Maybe Text

Response lenses

gesrBatchSize :: Lens' GetEventSourceResponse (Maybe Int)

The largest number of records that AWS Lambda will POST in the invocation request to your function.

gesrEventSource :: Lens' GetEventSourceResponse (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Kinesis stream that is the source of events.

gesrFunctionName :: Lens' GetEventSourceResponse (Maybe Text)

The Lambda function to invoke when AWS Lambda detects an event on the stream.

gesrIsActive :: Lens' GetEventSourceResponse (Maybe Bool)

Indicates whether the event source mapping is currently honored. Events are only processes if IsActive is true.

gesrLastModified :: Lens' GetEventSourceResponse (Maybe UTCTime)

The UTC time string indicating the last time the event mapping was updated.

gesrParameters :: Lens' GetEventSourceResponse (HashMap Text Text)

The map (key-value pairs) defining the configuration for AWS Lambda to use when reading the event source.

gesrRole :: Lens' GetEventSourceResponse (Maybe Text)

The ARN of the IAM role (invocation role) that AWS Lambda can assume to read from the stream and invoke the function.

gesrStatus :: Lens' GetEventSourceResponse (Maybe Text)

The description of the health of the event source mapping. Valid values are: PENDING, OK, and "PROBLEM:message". Initially this staus is PENDING. When AWS Lambda begins processing events, it changes the status to OK.

gesrUUID :: Lens' GetEventSourceResponse (Maybe Text)

The AWS Lambda assigned opaque identifier for the mapping.