Network.AWS.Lambda.GetEventSourceMapping

Description

Returns configuration information for the specified event source mapping (see CreateEventSourceMapping).

This operation requires permission for the 'lambda:GetEventSourceMapping' action.

http://docs.aws.amazon.com/lambda/latest/dg/API_GetEventSourceMapping.html

Request

data GetEventSourceMapping

Request constructor

getEventSourceMapping

Arguments

:: Text
gesmUUID

-> GetEventSourceMapping

GetEventSourceMapping constructor.

The fields accessible through corresponding lenses are:

gesmUUID :: Text

Request lenses

gesmUUID :: Lens' GetEventSourceMapping Text

The AWS Lambda assigned ID of the event source mapping.

Response

data GetEventSourceMappingResponse

Response constructor

getEventSourceMappingResponse :: GetEventSourceMappingResponse

GetEventSourceMappingResponse constructor.

The fields accessible through corresponding lenses are:

gesmrBatchSize :: Maybe Natural

gesmrEventSourceArn :: Maybe Text

gesmrFunctionArn :: Maybe Text

gesmrLastModified :: Maybe UTCTime

gesmrLastProcessingResult :: Maybe Text

gesmrState :: Maybe Text

gesmrStateTransitionReason :: Maybe Text

gesmrUUID :: Maybe Text

Response lenses

gesmrBatchSize :: Lens' GetEventSourceMappingResponse (Maybe Natural)

The largest number of records that AWS Lambda will retrieve from your event source at the time of invoking your function. Your function receives an event with all the retrieved records.

gesmrEventSourceArn :: Lens' GetEventSourceMappingResponse (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Kinesis stream that is the source of events.

gesmrFunctionArn :: Lens' GetEventSourceMappingResponse (Maybe Text)

The Lambda function to invoke when AWS Lambda detects an event on the stream.

gesmrLastModified :: Lens' GetEventSourceMappingResponse (Maybe UTCTime)

The UTC time string indicating the last time the event mapping was updated.

gesmrLastProcessingResult :: Lens' GetEventSourceMappingResponse (Maybe Text)

The result of the last AWS Lambda invocation of your Lambda function.

gesmrState :: Lens' GetEventSourceMappingResponse (Maybe Text)

The state of the event source mapping. It can be Creating, Enabled, Disabled, Enabling, Disabling, Updating, or Deleting.

gesmrStateTransitionReason :: Lens' GetEventSourceMappingResponse (Maybe Text)

The reason the event source mapping is in its current state. It is either user-requested or an AWS Lambda-initiated state transition.

gesmrUUID :: Lens' GetEventSourceMappingResponse (Maybe Text)

The AWS Lambda assigned opaque identifier for the mapping.