Network.AWS.Lambda.GetFunction

Description

Returns the configuration information of the Lambda function and a presigned URL link to the .zip file you uploaded with CreateFunction so you can download the .zip file. Note that the URL is valid for up to 10 minutes. The configuration information is the same information you provided as parameters when uploading the function.

This operation requires permission for the 'lambda:GetFunction' action.

http://docs.aws.amazon.com/lambda/latest/dg/API_GetFunction.html

Request

data GetFunction

Request constructor

getFunction

Arguments

:: Text
gfFunctionName

-> GetFunction

GetFunction constructor.

The fields accessible through corresponding lenses are:

gfFunctionName :: Text

Request lenses

gfFunctionName :: Lens' GetFunction Text

The Lambda function name.

You can specify an unqualified function name (for example, Thumbnail) or you can specify Amazon Resource Name (ARN) of the function (for example, "arn:aws:lambda:us-west-2:account-id:function:ThumbNail"). AWS Lambda also allows you to specify only the account ID qualifier (for example, "account-id:Thumbnail"). Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 character in length.

Response

data GetFunctionResponse

Response constructor

getFunctionResponse :: GetFunctionResponse

GetFunctionResponse constructor.

The fields accessible through corresponding lenses are:

gfrCode :: Maybe FunctionCodeLocation

gfrConfiguration :: Maybe FunctionConfiguration

Response lenses

gfrCode :: Lens' GetFunctionResponse (Maybe FunctionCodeLocation)

gfrConfiguration :: Lens' GetFunctionResponse (Maybe FunctionConfiguration)