Network.AWS.Lambda.GetFunctionConfiguration

Description

Returns the configuration information of the Lambda function. This the same information you provided as parameters when uploading the function by using CreateFunction.

This operation requires permission for the 'lambda:GetFunctionConfiguration' operation.

http://docs.aws.amazon.com/lambda/latest/dg/API_GetFunctionConfiguration.html

Request

data GetFunctionConfiguration

Request constructor

getFunctionConfiguration

Arguments

:: Text
gfcFunctionName

-> GetFunctionConfiguration

GetFunctionConfiguration constructor.

The fields accessible through corresponding lenses are:

gfcFunctionName :: Text

Request lenses

gfcFunctionName :: Lens' GetFunctionConfiguration Text

The name of the Lambda function for which you want to retrieve the configuration information.

You can specify an unqualified function name (for example, Thumbnail) or you can specify Amazon Resource Name (ARN) of the function (for example, "arn:aws:lambda:us-west-2:account-id:function:ThumbNail"). AWS Lambda also allows you to specify only the account ID qualifier (for example, "account-id:Thumbnail"). Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 character in length.

Response

data GetFunctionConfigurationResponse

Response constructor

getFunctionConfigurationResponse :: GetFunctionConfigurationResponse

GetFunctionConfigurationResponse constructor.

The fields accessible through corresponding lenses are:

gfcrCodeSize :: Maybe Integer

gfcrDescription :: Maybe Text

gfcrFunctionArn :: Maybe Text

gfcrFunctionName :: Maybe Text

gfcrHandler :: Maybe Text

gfcrLastModified :: Maybe Text

gfcrMemorySize :: Maybe Natural

gfcrRole :: Maybe Text

gfcrRuntime :: Maybe Runtime

gfcrTimeout :: Maybe Natural

Response lenses

gfcrCodeSize :: Lens' GetFunctionConfigurationResponse (Maybe Integer)

The size, in bytes, of the function .zip file you uploaded.

gfcrDescription :: Lens' GetFunctionConfigurationResponse (Maybe Text)

The user-provided description.

gfcrFunctionArn :: Lens' GetFunctionConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the function.

gfcrFunctionName :: Lens' GetFunctionConfigurationResponse (Maybe Text)

The name of the function.

gfcrHandler :: Lens' GetFunctionConfigurationResponse (Maybe Text)

The function Lambda calls to begin executing your function.

gfcrLastModified :: Lens' GetFunctionConfigurationResponse (Maybe Text)

The timestamp of the last time you updated the function.

gfcrMemorySize :: Lens' GetFunctionConfigurationResponse (Maybe Natural)

The memory size, in MB, you configured for the function. Must be a multiple of 64 MB.

gfcrRole :: Lens' GetFunctionConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources.

gfcrRuntime :: Lens' GetFunctionConfigurationResponse (Maybe Runtime)

The runtime environment for the Lambda function.

gfcrTimeout :: Lens' GetFunctionConfigurationResponse (Maybe Natural)

The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.