Network.AWS.Lambda.ListEventSourceMappings

Description

Returns a list of event source mappings you created using the CreateEventSourceMapping (see CreateEventSourceMapping), where you identify a stream as an event source. This list does not include Amazon S3 event sources.

For each mapping, the API returns configuration information. You can optionally specify filters to retrieve specific event source mappings.

This operation requires permission for the 'lambda:ListEventSourceMappings' action.

http://docs.aws.amazon.com/lambda/latest/dg/API_ListEventSourceMappings.html

Request

data ListEventSourceMappings

Request constructor

listEventSourceMappings :: ListEventSourceMappings

ListEventSourceMappings constructor.

The fields accessible through corresponding lenses are:

lesmEventSourceArn :: Maybe Text

lesmFunctionName :: Maybe Text

lesmMarker :: Maybe Text

lesmMaxItems :: Maybe Natural

Request lenses

lesmEventSourceArn :: Lens' ListEventSourceMappings (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Kinesis stream.

lesmFunctionName :: Lens' ListEventSourceMappings (Maybe Text)

The name of the Lambda function.

You can specify an unqualified function name (for example, Thumbnail) or you can specify Amazon Resource Name (ARN) of the function (for example, "arn:aws:lambda:us-west-2:account-id:function:ThumbNail"). AWS Lambda also allows you to specify only the account ID qualifier (for example, "account-id:Thumbnail"). Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 character in length.

lesmMarker :: Lens' ListEventSourceMappings (Maybe Text)

Optional string. An opaque pagination token returned from a previous ListEventSourceMappings operation. If present, specifies to continue the list from where the returning call left off.

lesmMaxItems :: Lens' ListEventSourceMappings (Maybe Natural)

Optional integer. Specifies the maximum number of event sources to return in response. This value must be greater than 0.

Response

data ListEventSourceMappingsResponse

Response constructor

listEventSourceMappingsResponse :: ListEventSourceMappingsResponse

ListEventSourceMappingsResponse constructor.

The fields accessible through corresponding lenses are:

lesmrEventSourceMappings :: [EventSourceMappingConfiguration]

lesmrNextMarker :: Maybe Text

Response lenses

lesmrEventSourceMappings :: Lens' ListEventSourceMappingsResponse [EventSourceMappingConfiguration]

An arrary of EventSourceMappingConfiguration objects.

lesmrNextMarker :: Lens' ListEventSourceMappingsResponse (Maybe Text)

A string, present if there are more event source mappings.