Network.AWS.Lambda.ListEventSources

Description

Returns a list of event source mappings. For each mapping, the API returns configuration information (see AddEventSource). You can optionally specify filters to retrieve specific event source mappings. This operation requires permission for the lambda:ListEventSources action.

http://docs.aws.amazon.com/lambda/latest/dg/API_ListEventSources.html

Request

data ListEventSources

Request constructor

listEventSources :: ListEventSources

ListEventSources constructor.

The fields accessible through corresponding lenses are:

lesEventSourceArn :: Maybe Text

lesFunctionName :: Maybe Text

lesMarker :: Maybe Text

lesMaxItems :: Maybe Natural

Request lenses

lesEventSourceArn :: Lens' ListEventSources (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Kinesis stream.

lesFunctionName :: Lens' ListEventSources (Maybe Text)

The name of the AWS Lambda function.

lesMarker :: Lens' ListEventSources (Maybe Text)

Optional string. An opaque pagination token returned from a previous ListEventSources operation. If present, specifies to continue the list from where the returning call left off.

lesMaxItems :: Lens' ListEventSources (Maybe Natural)

Optional integer. Specifies the maximum number of event sources to return in response. This value must be greater than 0.

Response

data ListEventSourcesResponse

Response constructor

listEventSourcesResponse :: ListEventSourcesResponse

ListEventSourcesResponse constructor.

The fields accessible through corresponding lenses are:

lesrEventSources :: [EventSourceConfiguration]

lesrNextMarker :: Maybe Text

Response lenses

lesrEventSources :: Lens' ListEventSourcesResponse [EventSourceConfiguration]

An arrary of EventSourceConfiguration objects.

lesrNextMarker :: Lens' ListEventSourcesResponse (Maybe Text)

A string, present if there are more event source mappings.