Network.AWS.Lambda.RemovePermission

Description

You can remove individual permissions from an access policy associated with a Lambda function by providing a Statement ID.

Note that removal of a permission will cause an active event source to lose permission to the function.

You need permission for the 'lambda:RemovePermission' action.

http://docs.aws.amazon.com/lambda/latest/dg/API_RemovePermission.html

Request

data RemovePermission

Request constructor

removePermission

Arguments

:: Text
rpFunctionName

-> Text
rpStatementId

-> RemovePermission

RemovePermission constructor.

The fields accessible through corresponding lenses are:

rpFunctionName :: Text

rpStatementId :: Text

Request lenses

rpFunctionName :: Lens' RemovePermission Text

Lambda function whose access policy you want to remove a permission from.

You can specify an unqualified function name (for example, Thumbnail) or you can specify Amazon Resource Name (ARN) of the function (for example, "arn:aws:lambda:us-west-2:account-id:function:ThumbNail"). AWS Lambda also allows you to specify only the account ID qualifier (for example, "account-id:Thumbnail"). Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 character in length.

rpStatementId :: Lens' RemovePermission Text

Statement ID of the permission to remove.

Response

data RemovePermissionResponse

Response constructor

removePermissionResponse :: RemovePermissionResponse

RemovePermissionResponse constructor.