Network.AWS.Lambda.TagResource

Description

Creates a list of tags (key-value pairs) on the Lambda function. Requires the Lambda function ARN (Amazon Resource Name). If a key is specified without a value, Lambda creates a tag with the specified key and a value of null.

Creating a Request

tagResource

Arguments

:: Text
trResource

-> TagResource

Creates a value of TagResource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

trResource - The ARN (Amazon Resource Name) of the Lambda function.

trTags - The list of tags (key-value pairs) you are assigning to the Lambda function.

data TagResource

See: tagResource smart constructor.

Request Lenses

trResource :: Lens' TagResource Text

The ARN (Amazon Resource Name) of the Lambda function.

trTags :: Lens' TagResource (HashMap Text Text)

The list of tags (key-value pairs) you are assigning to the Lambda function.

Destructuring the Response

tagResourceResponse :: TagResourceResponse

Creates a value of TagResourceResponse with the minimum fields required to make a request.

data TagResourceResponse

See: tagResourceResponse smart constructor.