Network.AWS.Lambda.Types

Service

data Lambda

Version 2014-11-11 of the Amazon Lambda service.

Error

data JSONError :: *

Runtime

data Runtime

Constructors

Nodejs
nodejs

Mode

data Mode

Constructors

Event
event

FunctionCodeLocation

data FunctionCodeLocation

functionCodeLocation :: FunctionCodeLocation

FunctionCodeLocation constructor.

The fields accessible through corresponding lenses are:

fclLocation :: Maybe Text

fclRepositoryType :: Maybe Text

fclLocation :: Lens' FunctionCodeLocation (Maybe Text)

The presigned URL you can use to download the function's .zip file that you previously uploaded. The URL is valid for up to 10 minutes.

fclRepositoryType :: Lens' FunctionCodeLocation (Maybe Text)

The repository from which you can download the function.

FunctionConfiguration

data FunctionConfiguration

functionConfiguration :: FunctionConfiguration

FunctionConfiguration constructor.

The fields accessible through corresponding lenses are:

fcCodeSize :: Maybe Integer

fcConfigurationId :: Maybe Text

fcDescription :: Maybe Text

fcFunctionARN :: Maybe Text

fcFunctionName :: Maybe Text

fcHandler :: Maybe Text

fcLastModified :: Maybe UTCTime

fcMemorySize :: Maybe Natural

fcMode :: Maybe Mode

fcRole :: Maybe Text

fcRuntime :: Maybe Runtime

fcTimeout :: Maybe Natural

fcCodeSize :: Lens' FunctionConfiguration (Maybe Integer)

The size, in bytes, of the function .zip file you uploaded.

fcConfigurationId :: Lens' FunctionConfiguration (Maybe Text)

A Lambda-assigned unique identifier for the current function code and related configuration.

fcDescription :: Lens' FunctionConfiguration (Maybe Text)

The user-provided description.

fcFunctionARN :: Lens' FunctionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) assigned to the function.

fcFunctionName :: Lens' FunctionConfiguration (Maybe Text)

The name of the function.

fcHandler :: Lens' FunctionConfiguration (Maybe Text)

The function Lambda calls to begin executing your function.

fcLastModified :: Lens' FunctionConfiguration (Maybe UTCTime)

The timestamp of the last time you updated the function.

fcMemorySize :: Lens' FunctionConfiguration (Maybe Natural)

The memory size, in MB, you configured for the function. Must be a multiple of 64 MB.

fcMode :: Lens' FunctionConfiguration (Maybe Mode)

The type of the Lambda function you uploaded.

fcRole :: Lens' FunctionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources.

fcRuntime :: Lens' FunctionConfiguration (Maybe Runtime)

The runtime environment for the Lambda function.

fcTimeout :: Lens' FunctionConfiguration (Maybe Natural)

The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

EventSourceConfiguration

data EventSourceConfiguration

eventSourceConfiguration :: EventSourceConfiguration

EventSourceConfiguration constructor.

The fields accessible through corresponding lenses are:

escBatchSize :: Maybe Int

escEventSource :: Maybe Text

escFunctionName :: Maybe Text

escIsActive :: Maybe Bool

escLastModified :: Maybe UTCTime

escParameters :: HashMap Text Text

escRole :: Maybe Text

escStatus :: Maybe Text

escUUID :: Maybe Text

escBatchSize :: Lens' EventSourceConfiguration (Maybe Int)

The largest number of records that AWS Lambda will POST in the invocation request to your function.

escEventSource :: Lens' EventSourceConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Kinesis stream that is the source of events.

escFunctionName :: Lens' EventSourceConfiguration (Maybe Text)

The Lambda function to invoke when AWS Lambda detects an event on the stream.

escIsActive :: Lens' EventSourceConfiguration (Maybe Bool)

Indicates whether the event source mapping is currently honored. Events are only processes if IsActive is true.

escLastModified :: Lens' EventSourceConfiguration (Maybe UTCTime)

The UTC time string indicating the last time the event mapping was updated.

escParameters :: Lens' EventSourceConfiguration (HashMap Text Text)

The map (key-value pairs) defining the configuration for AWS Lambda to use when reading the event source.

escRole :: Lens' EventSourceConfiguration (Maybe Text)

The ARN of the IAM role (invocation role) that AWS Lambda can assume to read from the stream and invoke the function.

escStatus :: Lens' EventSourceConfiguration (Maybe Text)

The description of the health of the event source mapping. Valid values are: PENDING, OK, and "PROBLEM:message". Initially this staus is PENDING. When AWS Lambda begins processing events, it changes the status to OK.

escUUID :: Lens' EventSourceConfiguration (Maybe Text)

The AWS Lambda assigned opaque identifier for the mapping.