Network.AWS.Lambda.UpdateAlias

Description

Using this API you can update the function version to which the alias points and the alias description. For more information, see Introduction to AWS Lambda Aliases .

This requires permission for the lambda:UpdateAlias action.

Creating a Request

updateAlias

Arguments

:: Text
uaFunctionName

-> Text
uaName

-> UpdateAlias

Creates a value of UpdateAlias with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uaFunctionVersion - Using this parameter you can change the Lambda function version to which the alias points.

uaDescription - You can change the description of the alias using this parameter.

uaFunctionName - The function name for which the alias is created. Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 characters in length.

uaName - The alias name.

data UpdateAlias

See: updateAlias smart constructor.

Request Lenses

uaFunctionVersion :: Lens' UpdateAlias (Maybe Text)

Using this parameter you can change the Lambda function version to which the alias points.

uaDescription :: Lens' UpdateAlias (Maybe Text)

You can change the description of the alias using this parameter.

uaFunctionName :: Lens' UpdateAlias Text

The function name for which the alias is created. Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 characters in length.

uaName :: Lens' UpdateAlias Text

The alias name.

Destructuring the Response

aliasConfiguration :: AliasConfiguration

Creates a value of AliasConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

acName - Alias name.

acFunctionVersion - Function version to which the alias points.

acAliasARN - Lambda function ARN that is qualified using the alias name as the suffix. For example, if you create an alias called BETA that points to a helloworld function version, the ARN is arn:aws:lambda:aws-regions:acct-id:function:helloworld:BETA .

acDescription - Alias description.

data AliasConfiguration

Provides configuration information about a Lambda function version alias.

See: aliasConfiguration smart constructor.

Response Lenses

acName :: Lens' AliasConfiguration (Maybe Text)

Alias name.

acFunctionVersion :: Lens' AliasConfiguration (Maybe Text)

Function version to which the alias points.

acAliasARN :: Lens' AliasConfiguration (Maybe Text)

Lambda function ARN that is qualified using the alias name as the suffix. For example, if you create an alias called BETA that points to a helloworld function version, the ARN is arn:aws:lambda:aws-regions:acct-id:function:helloworld:BETA .

acDescription :: Lens' AliasConfiguration (Maybe Text)

Alias description.