Network.AWS.Lambda.UpdateEventSourceMapping

Description

You can update an event source mapping. This is useful if you want to change the parameters of the existing mapping without losing your position in the stream. You can change which function will receive the stream records, but to change the stream itself, you must create a new mapping.

This operation requires permission for the 'lambda:UpdateEventSourceMapping' action.

http://docs.aws.amazon.com/lambda/latest/dg/API_UpdateEventSourceMapping.html

Request

data UpdateEventSourceMapping

Request constructor

updateEventSourceMapping

Arguments

:: Text
uesmUUID

-> UpdateEventSourceMapping

UpdateEventSourceMapping constructor.

The fields accessible through corresponding lenses are:

uesmBatchSize :: Maybe Natural

uesmEnabled :: Maybe Bool

uesmFunctionName :: Maybe Text

uesmUUID :: Text

Request lenses

uesmBatchSize :: Lens' UpdateEventSourceMapping (Maybe Natural)

The maximum number of stream records that can be sent to your Lambda function for a single invocation.

uesmEnabled :: Lens' UpdateEventSourceMapping (Maybe Bool)

Specifies whether AWS Lambda should actively poll the stream or not. If disabled, AWS Lambda will not poll the stream.

uesmFunctionName :: Lens' UpdateEventSourceMapping (Maybe Text)

The Lambda function to which you want the stream records sent.

You can specify an unqualified function name (for example, Thumbnail) or you can specify Amazon Resource Name (ARN) of the function (for example, "arn:aws:lambda:us-west-2:account-id:function:ThumbNail"). AWS Lambda also allows you to specify only the account ID qualifier (for example, "account-id:Thumbnail"). Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 character in length.

uesmUUID :: Lens' UpdateEventSourceMapping Text

The event source mapping identifier.

Response

data UpdateEventSourceMappingResponse

Response constructor

updateEventSourceMappingResponse :: UpdateEventSourceMappingResponse

UpdateEventSourceMappingResponse constructor.

The fields accessible through corresponding lenses are:

uesmrBatchSize :: Maybe Natural

uesmrEventSourceArn :: Maybe Text

uesmrFunctionArn :: Maybe Text

uesmrLastModified :: Maybe UTCTime

uesmrLastProcessingResult :: Maybe Text

uesmrState :: Maybe Text

uesmrStateTransitionReason :: Maybe Text

uesmrUUID :: Maybe Text

Response lenses

uesmrBatchSize :: Lens' UpdateEventSourceMappingResponse (Maybe Natural)

The largest number of records that AWS Lambda will retrieve from your event source at the time of invoking your function. Your function receives an event with all the retrieved records.

uesmrEventSourceArn :: Lens' UpdateEventSourceMappingResponse (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Kinesis stream that is the source of events.

uesmrFunctionArn :: Lens' UpdateEventSourceMappingResponse (Maybe Text)

The Lambda function to invoke when AWS Lambda detects an event on the stream.

uesmrLastModified :: Lens' UpdateEventSourceMappingResponse (Maybe UTCTime)

The UTC time string indicating the last time the event mapping was updated.

uesmrLastProcessingResult :: Lens' UpdateEventSourceMappingResponse (Maybe Text)

The result of the last AWS Lambda invocation of your Lambda function.

uesmrState :: Lens' UpdateEventSourceMappingResponse (Maybe Text)

The state of the event source mapping. It can be Creating, Enabled, Disabled, Enabling, Disabling, Updating, or Deleting.

uesmrStateTransitionReason :: Lens' UpdateEventSourceMappingResponse (Maybe Text)

The reason the event source mapping is in its current state. It is either user-requested or an AWS Lambda-initiated state transition.

uesmrUUID :: Lens' UpdateEventSourceMappingResponse (Maybe Text)

The AWS Lambda assigned opaque identifier for the mapping.