Network.AWS.Lambda.UpdateFunctionCode

Description

Updates the code for the specified Lambda function. This operation must only be used on an existing Lambda function and cannot be used to update the function configuration.

This operation requires permision for the 'lambda:UpdateFunctionCode' action.

http://docs.aws.amazon.com/lambda/latest/dg/API_UpdateFunctionCode.html

Request

data UpdateFunctionCode

Request constructor

updateFunctionCode

Arguments

:: Text
ufc1FunctionName

-> Base64
ufc1ZipFile

-> UpdateFunctionCode

UpdateFunctionCode constructor.

The fields accessible through corresponding lenses are:

ufc1FunctionName :: Text

ufc1ZipFile :: Base64

Request lenses

ufc1FunctionName :: Lens' UpdateFunctionCode Text

The existing Lambda function name whose code you want to replace.

You can specify an unqualified function name (for example, Thumbnail) or you can specify Amazon Resource Name (ARN) of the function (for example, "arn:aws:lambda:us-west-2:account-id:function:ThumbNail"). AWS Lambda also allows you to specify only the account ID qualifier (for example, "account-id:Thumbnail"). Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 character in length.

ufc1ZipFile :: Lens' UpdateFunctionCode Base64

Based64-encoded .zip file containing your packaged source code.

Response

data UpdateFunctionCodeResponse

Response constructor

updateFunctionCodeResponse :: UpdateFunctionCodeResponse

UpdateFunctionCodeResponse constructor.

The fields accessible through corresponding lenses are:

ufcrCodeSize :: Maybe Integer

ufcrDescription :: Maybe Text

ufcrFunctionArn :: Maybe Text

ufcrFunctionName :: Maybe Text

ufcrHandler :: Maybe Text

ufcrLastModified :: Maybe Text

ufcrMemorySize :: Maybe Natural

ufcrRole :: Maybe Text

ufcrRuntime :: Maybe Runtime

ufcrTimeout :: Maybe Natural

Response lenses

ufcrCodeSize :: Lens' UpdateFunctionCodeResponse (Maybe Integer)

The size, in bytes, of the function .zip file you uploaded.

ufcrDescription :: Lens' UpdateFunctionCodeResponse (Maybe Text)

The user-provided description.

ufcrFunctionArn :: Lens' UpdateFunctionCodeResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the function.

ufcrFunctionName :: Lens' UpdateFunctionCodeResponse (Maybe Text)

The name of the function.

ufcrHandler :: Lens' UpdateFunctionCodeResponse (Maybe Text)

The function Lambda calls to begin executing your function.

ufcrLastModified :: Lens' UpdateFunctionCodeResponse (Maybe Text)

The timestamp of the last time you updated the function.

ufcrMemorySize :: Lens' UpdateFunctionCodeResponse (Maybe Natural)

The memory size, in MB, you configured for the function. Must be a multiple of 64 MB.

ufcrRole :: Lens' UpdateFunctionCodeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources.

ufcrRuntime :: Lens' UpdateFunctionCodeResponse (Maybe Runtime)

The runtime environment for the Lambda function.

ufcrTimeout :: Lens' UpdateFunctionCodeResponse (Maybe Natural)

The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.