Network.AWS.Lambda.UpdateFunctionConfiguration

Description

Updates the configuration parameters for the specified Lambda function by using the values provided in the request. You provide only the parameters you want to change. This operation must only be used on an existing Lambda function and cannot be used to update the function's code.

This operation requires permission for the 'lambda:UpdateFunctionConfiguration' action.

http://docs.aws.amazon.com/lambda/latest/dg/API_UpdateFunctionConfiguration.html

Request

data UpdateFunctionConfiguration

Request constructor

updateFunctionConfiguration

Arguments

:: Text
ufcFunctionName

-> UpdateFunctionConfiguration

UpdateFunctionConfiguration constructor.

The fields accessible through corresponding lenses are:

ufcDescription :: Maybe Text

ufcFunctionName :: Text

ufcHandler :: Maybe Text

ufcMemorySize :: Maybe Natural

ufcRole :: Maybe Text

ufcTimeout :: Maybe Natural

Request lenses

ufcDescription :: Lens' UpdateFunctionConfiguration (Maybe Text)

A short user-defined function description. Lambda does not use this value. Assign a meaningful description as you see fit.

ufcFunctionName :: Lens' UpdateFunctionConfiguration Text

The name of the Lambda function.

ufcHandler :: Lens' UpdateFunctionConfiguration (Maybe Text)

The function that Lambda calls to begin executing your function. For Node.js, it is the module-name.export value in your function.

ufcMemorySize :: Lens' UpdateFunctionConfiguration (Maybe Natural)

The amount of memory, in MB, your Lambda function is given. Lambda uses this memory size to infer the amount of CPU allocated to your function. Your function use-case determines your CPU and memory requirements. For example, a database operation might need less memory compared to an image processing function. The default value is 128 MB. The value must be a multiple of 64 MB.

ufcRole :: Lens' UpdateFunctionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that Lambda will assume when it executes your function.

ufcTimeout :: Lens' UpdateFunctionConfiguration (Maybe Natural)

The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

Response

data UpdateFunctionConfigurationResponse

Response constructor

updateFunctionConfigurationResponse :: UpdateFunctionConfigurationResponse

UpdateFunctionConfigurationResponse constructor.

The fields accessible through corresponding lenses are:

ufcrCodeSize :: Maybe Integer

ufcrConfigurationId :: Maybe Text

ufcrDescription :: Maybe Text

ufcrFunctionARN :: Maybe Text

ufcrFunctionName :: Maybe Text

ufcrHandler :: Maybe Text

ufcrLastModified :: Maybe UTCTime

ufcrMemorySize :: Maybe Natural

ufcrMode :: Maybe Mode

ufcrRole :: Maybe Text

ufcrRuntime :: Maybe Runtime

ufcrTimeout :: Maybe Natural

Response lenses

ufcrCodeSize :: Lens' UpdateFunctionConfigurationResponse (Maybe Integer)

The size, in bytes, of the function .zip file you uploaded.

ufcrConfigurationId :: Lens' UpdateFunctionConfigurationResponse (Maybe Text)

A Lambda-assigned unique identifier for the current function code and related configuration.

ufcrDescription :: Lens' UpdateFunctionConfigurationResponse (Maybe Text)

The user-provided description.

ufcrFunctionARN :: Lens' UpdateFunctionConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the function.

ufcrFunctionName :: Lens' UpdateFunctionConfigurationResponse (Maybe Text)

The name of the function.

ufcrHandler :: Lens' UpdateFunctionConfigurationResponse (Maybe Text)

The function Lambda calls to begin executing your function.

ufcrLastModified :: Lens' UpdateFunctionConfigurationResponse (Maybe UTCTime)

The timestamp of the last time you updated the function.

ufcrMemorySize :: Lens' UpdateFunctionConfigurationResponse (Maybe Natural)

The memory size, in MB, you configured for the function. Must be a multiple of 64 MB.

ufcrMode :: Lens' UpdateFunctionConfigurationResponse (Maybe Mode)

The type of the Lambda function you uploaded.

ufcrRole :: Lens' UpdateFunctionConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources.

ufcrRuntime :: Lens' UpdateFunctionConfigurationResponse (Maybe Runtime)

The runtime environment for the Lambda function.

ufcrTimeout :: Lens' UpdateFunctionConfigurationResponse (Maybe Natural)

The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.