Network.AWS.Lambda

Description

AWS Lambda

Overview

This is the AWS Lambda API Reference . The AWS Lambda Developer Guide provides additional information. For the service overview, see What is AWS Lambda , and for information about how the service works, see AWS Lambda: How it Works in the AWS Lambda Developer Guide .

Service Configuration

lambda :: Service

API version 2015-03-31 of the Amazon Lambda SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Lambda.

KMSInvalidStateException

_KMSInvalidStateException :: AsError a => Getting (First ServiceError) a ServiceError

Lambda was unable to decrypt the environment variables because the KMS key used is in an invalid state for Decrypt. Check the function's KMS key settings.

EC2ThrottledException

_EC2ThrottledException :: AsError a => Getting (First ServiceError) a ServiceError

AWS Lambda was throttled by Amazon EC2 during Lambda function initialization using the execution role provided for the Lambda function.

InvalidRuntimeException

_InvalidRuntimeException :: AsError a => Getting (First ServiceError) a ServiceError

The runtime or runtime version specified is not supported.

PolicyLengthExceededException

_PolicyLengthExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Lambda function access policy is limited to 20 KB.

EC2AccessDeniedException

_EC2AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

InvalidSubnetIdException

_InvalidSubnetIdException :: AsError a => Getting (First ServiceError) a ServiceError

The Subnet ID provided in the Lambda function VPC configuration is invalid.

UnsupportedMediaTypeException

_UnsupportedMediaTypeException :: AsError a => Getting (First ServiceError) a ServiceError

The content type of the Invoke request body is not JSON.

InvalidRequestContentException

_InvalidRequestContentException :: AsError a => Getting (First ServiceError) a ServiceError

The request body could not be parsed as JSON.

KMSNotFoundException

_KMSNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Lambda was unable to decrypt the environment variables because the KMS key was not found. Check the function's KMS key settings.

ENILimitReachedException

_ENILimitReachedException :: AsError a => Getting (First ServiceError) a ServiceError

AWS Lambda was not able to create an Elastic Network Interface (ENI) in the VPC, specified as part of Lambda function configuration, because the limit for network interfaces has been reached.

InvalidParameterValueException

_InvalidParameterValueException :: AsError a => Getting (First ServiceError) a ServiceError

One of the parameters in the request is invalid. For example, if you provided an IAM role for AWS Lambda to assume in the CreateFunction or the UpdateFunctionConfiguration API, that AWS Lambda is unable to assume you will get this exception. You will also get this exception if you have selected a deprecated runtime, such as Node v0.10.42.

RequestTooLargeException

_RequestTooLargeException :: AsError a => Getting (First ServiceError) a ServiceError

The request payload exceeded the Invoke request body JSON input limit. For more information, see Limits .

TooManyRequestsException

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

InvalidSecurityGroupIdException

_InvalidSecurityGroupIdException :: AsError a => Getting (First ServiceError) a ServiceError

The Security Group ID provided in the Lambda function VPC configuration is invalid.

KMSDisabledException

_KMSDisabledException :: AsError a => Getting (First ServiceError) a ServiceError

Lambda was unable to decrypt the environment variables because the KMS key used is disabled. Check the Lambda function's KMS key settings.

SubnetIPAddressLimitReachedException

_SubnetIPAddressLimitReachedException :: AsError a => Getting (First ServiceError) a ServiceError

AWS Lambda was not able to set up VPC access for the Lambda function because one or more configured subnets has no available IP addresses.

ServiceException

_ServiceException :: AsError a => Getting (First ServiceError) a ServiceError

The AWS Lambda service encountered an internal error.

CodeStorageExceededException

_CodeStorageExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You have exceeded your maximum total code size per account. Limits

InvalidZipFileException

_InvalidZipFileException :: AsError a => Getting (First ServiceError) a ServiceError

AWS Lambda could not unzip the function zip file.

ResourceConflictException

_ResourceConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The resource already exists.

EC2UnexpectedException

_EC2UnexpectedException :: AsError a => Getting (First ServiceError) a ServiceError

AWS Lambda received an unexpected EC2 client exception while setting up for the Lambda function.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resource (for example, a Lambda function or access policy statement) specified in the request does not exist.

KMSAccessDeniedException

_KMSAccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

Lambda was unable to decrypt the environment variables because KMS access was denied. Check the Lambda function's KMS permissions.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetFunctionConfiguration

module Network.AWS.Lambda.GetFunctionConfiguration

DeleteEventSourceMapping

module Network.AWS.Lambda.DeleteEventSourceMapping

UpdateEventSourceMapping

module Network.AWS.Lambda.UpdateEventSourceMapping

CreateAlias

module Network.AWS.Lambda.CreateAlias

ListVersionsByFunction

module Network.AWS.Lambda.ListVersionsByFunction

ListAliases

module Network.AWS.Lambda.ListAliases

RemovePermission

module Network.AWS.Lambda.RemovePermission

Invoke

module Network.AWS.Lambda.Invoke

GetAlias

module Network.AWS.Lambda.GetAlias

GetEventSourceMapping

module Network.AWS.Lambda.GetEventSourceMapping

CreateFunction

module Network.AWS.Lambda.CreateFunction

CreateEventSourceMapping

module Network.AWS.Lambda.CreateEventSourceMapping

GetFunction

module Network.AWS.Lambda.GetFunction

ListEventSourceMappings (Paginated)

module Network.AWS.Lambda.ListEventSourceMappings

DeleteAlias

module Network.AWS.Lambda.DeleteAlias

UpdateAlias

module Network.AWS.Lambda.UpdateAlias

GetAccountSettings

module Network.AWS.Lambda.GetAccountSettings

AddPermission

module Network.AWS.Lambda.AddPermission

TagResource

module Network.AWS.Lambda.TagResource

PublishVersion

module Network.AWS.Lambda.PublishVersion

ListTags

module Network.AWS.Lambda.ListTags

DeleteFunction

module Network.AWS.Lambda.DeleteFunction

UntagResource

module Network.AWS.Lambda.UntagResource

UpdateFunctionConfiguration

module Network.AWS.Lambda.UpdateFunctionConfiguration

ListFunctions (Paginated)

module Network.AWS.Lambda.ListFunctions

UpdateFunctionCode

module Network.AWS.Lambda.UpdateFunctionCode

GetPolicy

module Network.AWS.Lambda.GetPolicy

Types

EventSourcePosition

data EventSourcePosition

Constructors

AtTimestamp

Latest

TrimHorizon

FunctionVersion

data FunctionVersion

Constructors

All

InvocationType

data InvocationType

Constructors

DryRun

Event

RequestResponse

LogType

data LogType

Constructors

None

Tail

Runtime

data Runtime

Constructors

DOTNETCORE1_0

JAVA8

NODEJS4_3

NODEJS4_3Edge

NODEJS6_10

Nodejs

PYTHON2_7

PYTHON3_6

TracingMode

data TracingMode

Constructors

Active

PassThrough

AccountLimit

data AccountLimit

Provides limits of code size and concurrency associated with the current account and region.

See: accountLimit smart constructor.

accountLimit :: AccountLimit

Creates a value of AccountLimit with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

alConcurrentExecutions - Number of simultaneous executions of your function per region. For more information or to request a limit increase for concurrent executions, see Lambda Function Concurrent Executions . The default limit is 100.

alTotalCodeSize - Maximum size, in bytes, of a code package you can upload per region. The default size is 75 GB.

alCodeSizeUnzipped - Size, in bytes, of codedependencies that you can zip into a deployment package (uncompressed zipjar size) for uploading. The default limit is 250 MB.

alCodeSizeZipped - Size, in bytes, of a single zipped codedependencies package you can upload for your Lambda function(.zip.jar file). Try using Amazon S3 for uploading larger files. Default limit is 50 MB.

alConcurrentExecutions :: Lens' AccountLimit (Maybe Int)

Number of simultaneous executions of your function per region. For more information or to request a limit increase for concurrent executions, see Lambda Function Concurrent Executions . The default limit is 100.

alTotalCodeSize :: Lens' AccountLimit (Maybe Integer)

Maximum size, in bytes, of a code package you can upload per region. The default size is 75 GB.

alCodeSizeUnzipped :: Lens' AccountLimit (Maybe Integer)

Size, in bytes, of codedependencies that you can zip into a deployment package (uncompressed zipjar size) for uploading. The default limit is 250 MB.

alCodeSizeZipped :: Lens' AccountLimit (Maybe Integer)

Size, in bytes, of a single zipped codedependencies package you can upload for your Lambda function(.zip.jar file). Try using Amazon S3 for uploading larger files. Default limit is 50 MB.

AccountUsage

data AccountUsage

Provides code size usage and function count associated with the current account and region.

See: accountUsage smart constructor.

accountUsage :: AccountUsage

Creates a value of AccountUsage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

auTotalCodeSize - Total size, in bytes, of the account's deployment packages per region.

auFunctionCount - The number of your account's existing functions per region.

auTotalCodeSize :: Lens' AccountUsage (Maybe Integer)

Total size, in bytes, of the account's deployment packages per region.

auFunctionCount :: Lens' AccountUsage (Maybe Integer)

The number of your account's existing functions per region.

AliasConfiguration

data AliasConfiguration

Provides configuration information about a Lambda function version alias.

See: aliasConfiguration smart constructor.

aliasConfiguration :: AliasConfiguration

Creates a value of AliasConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

acName - Alias name.

acFunctionVersion - Function version to which the alias points.

acAliasARN - Lambda function ARN that is qualified using the alias name as the suffix. For example, if you create an alias called BETA that points to a helloworld function version, the ARN is arn:aws:lambda:aws-regions:acct-id:function:helloworld:BETA .

acDescription - Alias description.

acName :: Lens' AliasConfiguration (Maybe Text)

Alias name.

acFunctionVersion :: Lens' AliasConfiguration (Maybe Text)

Function version to which the alias points.

acAliasARN :: Lens' AliasConfiguration (Maybe Text)

Lambda function ARN that is qualified using the alias name as the suffix. For example, if you create an alias called BETA that points to a helloworld function version, the ARN is arn:aws:lambda:aws-regions:acct-id:function:helloworld:BETA .

acDescription :: Lens' AliasConfiguration (Maybe Text)

Alias description.

DeadLetterConfig

data DeadLetterConfig

The parent object that contains the target ARN (Amazon Resource Name) of an Amazon SQS queue or Amazon SNS topic.

See: deadLetterConfig smart constructor.

deadLetterConfig :: DeadLetterConfig

Creates a value of DeadLetterConfig with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlcTargetARN - The Amazon Resource Name (ARN) of an Amazon SQS queue or Amazon SNS topic you specify as your Dead Letter Queue (DLQ).

dlcTargetARN :: Lens' DeadLetterConfig (Maybe Text)

The Amazon Resource Name (ARN) of an Amazon SQS queue or Amazon SNS topic you specify as your Dead Letter Queue (DLQ).

Environment

data Environment

The parent object that contains your environment's configuration settings.

See: environment smart constructor.

environment :: Environment

Creates a value of Environment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eVariables - The key-value pairs that represent your environment's configuration settings.

eVariables :: Lens' Environment (Maybe (HashMap Text Text))

The key-value pairs that represent your environment's configuration settings.

EnvironmentError

data EnvironmentError

The parent object that contains error information associated with your configuration settings.

See: environmentError smart constructor.

environmentError :: EnvironmentError

Creates a value of EnvironmentError with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eeErrorCode - The error code returned by the environment error object.

eeMessage - The message returned by the environment error object.

eeErrorCode :: Lens' EnvironmentError (Maybe Text)

The error code returned by the environment error object.

eeMessage :: Lens' EnvironmentError (Maybe Text)

The message returned by the environment error object.

EnvironmentResponse

data EnvironmentResponse

The parent object returned that contains your environment's configuration settings or any error information associated with your configuration settings.

See: environmentResponse smart constructor.

environmentResponse :: EnvironmentResponse

Creates a value of EnvironmentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

envVariables - The key-value pairs returned that represent your environment's configuration settings or error information.

envError - Undocumented member.

envVariables :: Lens' EnvironmentResponse (Maybe (HashMap Text Text))

The key-value pairs returned that represent your environment's configuration settings or error information.

envError :: Lens' EnvironmentResponse (Maybe EnvironmentError)

Undocumented member.

EventSourceMappingConfiguration

data EventSourceMappingConfiguration

Describes mapping between an Amazon Kinesis stream and a Lambda function.

See: eventSourceMappingConfiguration smart constructor.

eventSourceMappingConfiguration :: EventSourceMappingConfiguration

Creates a value of EventSourceMappingConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

esmcEventSourceARN - The Amazon Resource Name (ARN) of the Amazon Kinesis stream that is the source of events.

esmcState - The state of the event source mapping. It can be Creating , Enabled , Disabled , Enabling , Disabling , Updating , or Deleting .

esmcFunctionARN - The Lambda function to invoke when AWS Lambda detects an event on the stream.

esmcUUId - The AWS Lambda assigned opaque identifier for the mapping.

esmcLastProcessingResult - The result of the last AWS Lambda invocation of your Lambda function.

esmcBatchSize - The largest number of records that AWS Lambda will retrieve from your event source at the time of invoking your function. Your function receives an event with all the retrieved records.

esmcStateTransitionReason - The reason the event source mapping is in its current state. It is either user-requested or an AWS Lambda-initiated state transition.

esmcLastModified - The UTC time string indicating the last time the event mapping was updated.

esmcEventSourceARN :: Lens' EventSourceMappingConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Kinesis stream that is the source of events.

esmcState :: Lens' EventSourceMappingConfiguration (Maybe Text)

The state of the event source mapping. It can be Creating , Enabled , Disabled , Enabling , Disabling , Updating , or Deleting .

esmcFunctionARN :: Lens' EventSourceMappingConfiguration (Maybe Text)

The Lambda function to invoke when AWS Lambda detects an event on the stream.

esmcUUId :: Lens' EventSourceMappingConfiguration (Maybe Text)

The AWS Lambda assigned opaque identifier for the mapping.

esmcLastProcessingResult :: Lens' EventSourceMappingConfiguration (Maybe Text)

The result of the last AWS Lambda invocation of your Lambda function.

esmcBatchSize :: Lens' EventSourceMappingConfiguration (Maybe Natural)

The largest number of records that AWS Lambda will retrieve from your event source at the time of invoking your function. Your function receives an event with all the retrieved records.

esmcStateTransitionReason :: Lens' EventSourceMappingConfiguration (Maybe Text)

The reason the event source mapping is in its current state. It is either user-requested or an AWS Lambda-initiated state transition.

esmcLastModified :: Lens' EventSourceMappingConfiguration (Maybe UTCTime)

The UTC time string indicating the last time the event mapping was updated.

FunctionCode

data FunctionCode

The code for the Lambda function.

See: functionCode smart constructor.

functionCode :: FunctionCode

Creates a value of FunctionCode with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fcS3ObjectVersion - The Amazon S3 object (the deployment package) version you want to upload.

fcS3Key - The Amazon S3 object (the deployment package) key name you want to upload.

fcZipFile - The contents of your zip file containing your deployment package. If you are using the web API directly, the contents of the zip file must be base64-encoded. If you are using the AWS SDKs or the AWS CLI, the SDKs or CLI will do the encoding for you. For more information about creating a .zip file, see Execution Permissions in the AWS Lambda Developer Guide . -- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

fcS3Bucket - Amazon S3 bucket name where the .zip file containing your deployment package is stored. This bucket must reside in the same AWS region where you are creating the Lambda function.

fcS3ObjectVersion :: Lens' FunctionCode (Maybe Text)

The Amazon S3 object (the deployment package) version you want to upload.

fcS3Key :: Lens' FunctionCode (Maybe Text)

The Amazon S3 object (the deployment package) key name you want to upload.

fcZipFile :: Lens' FunctionCode (Maybe ByteString)

The contents of your zip file containing your deployment package. If you are using the web API directly, the contents of the zip file must be base64-encoded. If you are using the AWS SDKs or the AWS CLI, the SDKs or CLI will do the encoding for you. For more information about creating a .zip file, see Execution Permissions in the AWS Lambda Developer Guide . -- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

fcS3Bucket :: Lens' FunctionCode (Maybe Text)

Amazon S3 bucket name where the .zip file containing your deployment package is stored. This bucket must reside in the same AWS region where you are creating the Lambda function.

FunctionCodeLocation

data FunctionCodeLocation

The object for the Lambda function location.

See: functionCodeLocation smart constructor.

functionCodeLocation :: FunctionCodeLocation

Creates a value of FunctionCodeLocation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fclLocation - The presigned URL you can use to download the function's .zip file that you previously uploaded. The URL is valid for up to 10 minutes.

fclRepositoryType - The repository from which you can download the function.

fclLocation :: Lens' FunctionCodeLocation (Maybe Text)

The presigned URL you can use to download the function's .zip file that you previously uploaded. The URL is valid for up to 10 minutes.

fclRepositoryType :: Lens' FunctionCodeLocation (Maybe Text)

The repository from which you can download the function.

FunctionConfiguration

data FunctionConfiguration

A complex type that describes function metadata.

See: functionConfiguration smart constructor.

functionConfiguration :: FunctionConfiguration

Creates a value of FunctionConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fcMemorySize - The memory size, in MB, you configured for the function. Must be a multiple of 64 MB.

fcRuntime - The runtime environment for the Lambda function.

fcFunctionARN - The Amazon Resource Name (ARN) assigned to the function.

fcKMSKeyARN - The Amazon Resource Name (ARN) of the KMS key used to encrypt your function's environment variables. If empty, it means you are using the AWS Lambda default service key.

fcEnvironment - The parent object that contains your environment's configuration settings.

fcDeadLetterConfig - The parent object that contains the target ARN (Amazon Resource Name) of an Amazon SQS queue or Amazon SNS topic.

fcRole - The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources.

fcVPCConfig - VPC configuration associated with your Lambda function.

fcVersion - The version of the Lambda function.

fcFunctionName - The name of the function. Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 characters in length.

fcCodeSize - The size, in bytes, of the function .zip file you uploaded.

fcHandler - The function Lambda calls to begin executing your function.

fcTimeout - The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

fcLastModified - The time stamp of the last time you updated the function. The time stamp is conveyed as a string complying with ISO-8601 in this way YYYY-MM-DDThh:mm:ssTZD (e.g., 1997-07-16T19:20:30+01:00). For more information, see Date and Time Formats .

fcCodeSha256 - It is the SHA256 hash of your function deployment package.

fcTracingConfig - The parent object that contains your function's tracing settings.

fcDescription - The user-provided description.

fcMasterARN - Returns the ARN (Amazon Resource Name) of the master function.

fcMemorySize :: Lens' FunctionConfiguration (Maybe Natural)

The memory size, in MB, you configured for the function. Must be a multiple of 64 MB.

fcRuntime :: Lens' FunctionConfiguration (Maybe Runtime)

The runtime environment for the Lambda function.

fcFunctionARN :: Lens' FunctionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) assigned to the function.

fcKMSKeyARN :: Lens' FunctionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the KMS key used to encrypt your function's environment variables. If empty, it means you are using the AWS Lambda default service key.

fcEnvironment :: Lens' FunctionConfiguration (Maybe EnvironmentResponse)

The parent object that contains your environment's configuration settings.

fcDeadLetterConfig :: Lens' FunctionConfiguration (Maybe DeadLetterConfig)

The parent object that contains the target ARN (Amazon Resource Name) of an Amazon SQS queue or Amazon SNS topic.

fcRole :: Lens' FunctionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources.

fcVPCConfig :: Lens' FunctionConfiguration (Maybe VPCConfigResponse)

VPC configuration associated with your Lambda function.

fcVersion :: Lens' FunctionConfiguration (Maybe Text)

The version of the Lambda function.

fcFunctionName :: Lens' FunctionConfiguration (Maybe Text)

The name of the function. Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 characters in length.

fcCodeSize :: Lens' FunctionConfiguration (Maybe Integer)

The size, in bytes, of the function .zip file you uploaded.

fcHandler :: Lens' FunctionConfiguration (Maybe Text)

The function Lambda calls to begin executing your function.

fcTimeout :: Lens' FunctionConfiguration (Maybe Natural)

The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

fcLastModified :: Lens' FunctionConfiguration (Maybe Text)

The time stamp of the last time you updated the function. The time stamp is conveyed as a string complying with ISO-8601 in this way YYYY-MM-DDThh:mm:ssTZD (e.g., 1997-07-16T19:20:30+01:00). For more information, see Date and Time Formats .

fcCodeSha256 :: Lens' FunctionConfiguration (Maybe Text)

It is the SHA256 hash of your function deployment package.

fcTracingConfig :: Lens' FunctionConfiguration (Maybe TracingConfigResponse)

The parent object that contains your function's tracing settings.

fcDescription :: Lens' FunctionConfiguration (Maybe Text)

The user-provided description.

fcMasterARN :: Lens' FunctionConfiguration (Maybe Text)

Returns the ARN (Amazon Resource Name) of the master function.

TracingConfig

data TracingConfig

The parent object that contains your function's tracing settings.

See: tracingConfig smart constructor.

tracingConfig :: TracingConfig

Creates a value of TracingConfig with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tMode - Can be either PassThrough or Active. If PassThrough, Lambda will only trace the request from an upstream service if it contains a tracing header with "sampled=1". If Active, Lambda will respect any tracing header it receives from an upstream service. If no tracing header is received, Lambda will call X-Ray for a tracing decision.

tMode :: Lens' TracingConfig (Maybe TracingMode)

Can be either PassThrough or Active. If PassThrough, Lambda will only trace the request from an upstream service if it contains a tracing header with "sampled=1". If Active, Lambda will respect any tracing header it receives from an upstream service. If no tracing header is received, Lambda will call X-Ray for a tracing decision.

TracingConfigResponse

data TracingConfigResponse

Parent object of the tracing information associated with your Lambda function.

See: tracingConfigResponse smart constructor.

tracingConfigResponse :: TracingConfigResponse

Creates a value of TracingConfigResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tcMode - The tracing mode associated with your Lambda function.

tcMode :: Lens' TracingConfigResponse (Maybe TracingMode)

The tracing mode associated with your Lambda function.

VPCConfig

data VPCConfig

If your Lambda function accesses resources in a VPC, you provide this parameter identifying the list of security group IDs and subnet IDs. These must belong to the same VPC. You must provide at least one security group and one subnet ID.

See: vpcConfig smart constructor.

vpcConfig :: VPCConfig

Creates a value of VPCConfig with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vpccSecurityGroupIds - A list of one or more security groups IDs in your VPC.

vpccSubnetIds - A list of one or more subnet IDs in your VPC.

vpccSecurityGroupIds :: Lens' VPCConfig [Text]

A list of one or more security groups IDs in your VPC.

vpccSubnetIds :: Lens' VPCConfig [Text]

A list of one or more subnet IDs in your VPC.

VPCConfigResponse

data VPCConfigResponse

VPC configuration associated with your Lambda function.

See: vpcConfigResponse smart constructor.

vpcConfigResponse :: VPCConfigResponse

Creates a value of VPCConfigResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vcSecurityGroupIds - A list of security group IDs associated with the Lambda function.

vcSubnetIds - A list of subnet IDs associated with the Lambda function.

vcVPCId - The VPC ID associated with you Lambda function.

vcSecurityGroupIds :: Lens' VPCConfigResponse [Text]

A list of security group IDs associated with the Lambda function.

vcSubnetIds :: Lens' VPCConfigResponse [Text]

A list of subnet IDs associated with the Lambda function.

vcVPCId :: Lens' VPCConfigResponse (Maybe Text)

The VPC ID associated with you Lambda function.