Network.AWS.Lambda.CreateFunction

Description

Creates a new Lambda function. The function metadata is created from the request parameters, and the code for the function is provided by a .zip file in the request body. If the function name already exists, the operation will fail. Note that the function name is case-sensitive.

If you are using versioning, you can also publish a version of the Lambda function you are creating using the Publish parameter. For more information about versioning, see AWS Lambda Function Versioning and Aliases .

This operation requires permission for the lambda:CreateFunction action.

Creating a Request

createFunction

Arguments

:: Text
cfFunctionName

-> Runtime
cfRuntime

-> Text
cfRole

-> Text
cfHandler

-> FunctionCode
cfCode

-> CreateFunction

Creates a value of CreateFunction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cfMemorySize - The amount of memory, in MB, your Lambda function is given. Lambda uses this memory size to infer the amount of CPU and memory allocated to your function. Your function use-case determines your CPU and memory requirements. For example, a database operation might need less memory compared to an image processing function. The default value is 128 MB. The value must be a multiple of 64 MB.

cfKMSKeyARN - The Amazon Resource Name (ARN) of the KMS key used to encrypt your function's environment variables. If not provided, AWS Lambda will use a default service key.

cfEnvironment - Undocumented member.

cfDeadLetterConfig - The parent object that contains the target ARN (Amazon Resource Name) of an Amazon SQS queue or Amazon SNS topic.

cfVPCConfig - If your Lambda function accesses resources in a VPC, you provide this parameter identifying the list of security group IDs and subnet IDs. These must belong to the same VPC. You must provide at least one security group and one subnet ID.

cfTimeout - The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

cfTracingConfig - The parent object that contains your function's tracing settings.

cfDescription - A short, user-defined function description. Lambda does not use this value. Assign a meaningful description as you see fit.

cfTags - The list of tags (key-value pairs) assigned to the new function.

cfPublish - This boolean parameter can be used to request AWS Lambda to create the Lambda function and publish a version as an atomic operation.

cfFunctionName - The name you want to assign to the function you are uploading. The function names appear in the console and are returned in the ListFunctions API. Function names are used to specify functions to other AWS Lambda API operations, such as Invoke . Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 characters in length.

cfRuntime - The runtime environment for the Lambda function you are uploading. To use the Python runtime v3.6, set the value to "python3.6". To use the Python runtime v2.7, set the value to "python2.7". To use the Node.js runtime v6.10, set the value to "nodejs6.10". To use the Node.js runtime v4.3, set the value to "nodejs4.3".

cfRole - The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources. For more information, see AWS Lambda: How it Works .

cfHandler - The function within your code that Lambda calls to begin execution. For Node.js, it is the module-name .export value in your function. For Java, it can be package.class-name::handler or package.class-name . For more information, see Lambda Function Handler (Java) .

cfCode - The code for the Lambda function.

data CreateFunction

See: createFunction smart constructor.

Request Lenses

cfMemorySize :: Lens' CreateFunction (Maybe Natural)

The amount of memory, in MB, your Lambda function is given. Lambda uses this memory size to infer the amount of CPU and memory allocated to your function. Your function use-case determines your CPU and memory requirements. For example, a database operation might need less memory compared to an image processing function. The default value is 128 MB. The value must be a multiple of 64 MB.

cfKMSKeyARN :: Lens' CreateFunction (Maybe Text)

The Amazon Resource Name (ARN) of the KMS key used to encrypt your function's environment variables. If not provided, AWS Lambda will use a default service key.

cfEnvironment :: Lens' CreateFunction (Maybe Environment)

Undocumented member.

cfDeadLetterConfig :: Lens' CreateFunction (Maybe DeadLetterConfig)

The parent object that contains the target ARN (Amazon Resource Name) of an Amazon SQS queue or Amazon SNS topic.

cfVPCConfig :: Lens' CreateFunction (Maybe VPCConfig)

If your Lambda function accesses resources in a VPC, you provide this parameter identifying the list of security group IDs and subnet IDs. These must belong to the same VPC. You must provide at least one security group and one subnet ID.

cfTimeout :: Lens' CreateFunction (Maybe Natural)

The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

cfTracingConfig :: Lens' CreateFunction (Maybe TracingConfig)

The parent object that contains your function's tracing settings.

cfDescription :: Lens' CreateFunction (Maybe Text)

A short, user-defined function description. Lambda does not use this value. Assign a meaningful description as you see fit.

cfTags :: Lens' CreateFunction (HashMap Text Text)

The list of tags (key-value pairs) assigned to the new function.

cfPublish :: Lens' CreateFunction (Maybe Bool)

This boolean parameter can be used to request AWS Lambda to create the Lambda function and publish a version as an atomic operation.

cfFunctionName :: Lens' CreateFunction Text

The name you want to assign to the function you are uploading. The function names appear in the console and are returned in the ListFunctions API. Function names are used to specify functions to other AWS Lambda API operations, such as Invoke . Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 characters in length.

cfRuntime :: Lens' CreateFunction Runtime

The runtime environment for the Lambda function you are uploading. To use the Python runtime v3.6, set the value to "python3.6". To use the Python runtime v2.7, set the value to "python2.7". To use the Node.js runtime v6.10, set the value to "nodejs6.10". To use the Node.js runtime v4.3, set the value to "nodejs4.3".

cfRole :: Lens' CreateFunction Text

The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources. For more information, see AWS Lambda: How it Works .

cfHandler :: Lens' CreateFunction Text

The function within your code that Lambda calls to begin execution. For Node.js, it is the module-name .export value in your function. For Java, it can be package.class-name::handler or package.class-name . For more information, see Lambda Function Handler (Java) .

cfCode :: Lens' CreateFunction FunctionCode

The code for the Lambda function.

Destructuring the Response

functionConfiguration :: FunctionConfiguration

Creates a value of FunctionConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fcMemorySize - The memory size, in MB, you configured for the function. Must be a multiple of 64 MB.

fcRuntime - The runtime environment for the Lambda function.

fcFunctionARN - The Amazon Resource Name (ARN) assigned to the function.

fcKMSKeyARN - The Amazon Resource Name (ARN) of the KMS key used to encrypt your function's environment variables. If empty, it means you are using the AWS Lambda default service key.

fcEnvironment - The parent object that contains your environment's configuration settings.

fcDeadLetterConfig - The parent object that contains the target ARN (Amazon Resource Name) of an Amazon SQS queue or Amazon SNS topic.

fcRole - The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources.

fcVPCConfig - VPC configuration associated with your Lambda function.

fcVersion - The version of the Lambda function.

fcFunctionName - The name of the function. Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 characters in length.

fcCodeSize - The size, in bytes, of the function .zip file you uploaded.

fcHandler - The function Lambda calls to begin executing your function.

fcTimeout - The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

fcLastModified - The time stamp of the last time you updated the function. The time stamp is conveyed as a string complying with ISO-8601 in this way YYYY-MM-DDThh:mm:ssTZD (e.g., 1997-07-16T19:20:30+01:00). For more information, see Date and Time Formats .

fcCodeSha256 - It is the SHA256 hash of your function deployment package.

fcTracingConfig - The parent object that contains your function's tracing settings.

fcDescription - The user-provided description.

fcMasterARN - Returns the ARN (Amazon Resource Name) of the master function.

data FunctionConfiguration

A complex type that describes function metadata.

See: functionConfiguration smart constructor.

Response Lenses

fcMemorySize :: Lens' FunctionConfiguration (Maybe Natural)

The memory size, in MB, you configured for the function. Must be a multiple of 64 MB.

fcRuntime :: Lens' FunctionConfiguration (Maybe Runtime)

The runtime environment for the Lambda function.

fcFunctionARN :: Lens' FunctionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) assigned to the function.

fcKMSKeyARN :: Lens' FunctionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the KMS key used to encrypt your function's environment variables. If empty, it means you are using the AWS Lambda default service key.

fcEnvironment :: Lens' FunctionConfiguration (Maybe EnvironmentResponse)

The parent object that contains your environment's configuration settings.

fcDeadLetterConfig :: Lens' FunctionConfiguration (Maybe DeadLetterConfig)

The parent object that contains the target ARN (Amazon Resource Name) of an Amazon SQS queue or Amazon SNS topic.

fcRole :: Lens' FunctionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources.

fcVPCConfig :: Lens' FunctionConfiguration (Maybe VPCConfigResponse)

VPC configuration associated with your Lambda function.

fcVersion :: Lens' FunctionConfiguration (Maybe Text)

The version of the Lambda function.

fcFunctionName :: Lens' FunctionConfiguration (Maybe Text)

The name of the function. Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 characters in length.

fcCodeSize :: Lens' FunctionConfiguration (Maybe Integer)

The size, in bytes, of the function .zip file you uploaded.

fcHandler :: Lens' FunctionConfiguration (Maybe Text)

The function Lambda calls to begin executing your function.

fcTimeout :: Lens' FunctionConfiguration (Maybe Natural)

The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

fcLastModified :: Lens' FunctionConfiguration (Maybe Text)

The time stamp of the last time you updated the function. The time stamp is conveyed as a string complying with ISO-8601 in this way YYYY-MM-DDThh:mm:ssTZD (e.g., 1997-07-16T19:20:30+01:00). For more information, see Date and Time Formats .

fcCodeSha256 :: Lens' FunctionConfiguration (Maybe Text)

It is the SHA256 hash of your function deployment package.

fcTracingConfig :: Lens' FunctionConfiguration (Maybe TracingConfigResponse)

The parent object that contains your function's tracing settings.

fcDescription :: Lens' FunctionConfiguration (Maybe Text)

The user-provided description.

fcMasterARN :: Lens' FunctionConfiguration (Maybe Text)

Returns the ARN (Amazon Resource Name) of the master function.