Network.AWS.Lambda.ListEventSourceMappings

Description

Returns a list of event source mappings you created using the CreateEventSourceMapping (see CreateEventSourceMapping ).

For each mapping, the API returns configuration information. You can optionally specify filters to retrieve specific event source mappings.

If you are using the versioning feature, you can get list of event source mappings for a specific Lambda function version or an alias as described in the FunctionName parameter. For information about the versioning feature, see AWS Lambda Function Versioning and Aliases .

This operation requires permission for the lambda:ListEventSourceMappings action.

This operation returns paginated results.

Creating a Request

listEventSourceMappings :: ListEventSourceMappings

Creates a value of ListEventSourceMappings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lesmEventSourceARN - The Amazon Resource Name (ARN) of the Amazon Kinesis stream. (This parameter is optional.)

lesmMarker - Optional string. An opaque pagination token returned from a previous ListEventSourceMappings operation. If present, specifies to continue the list from where the returning call left off.

lesmMaxItems - Optional integer. Specifies the maximum number of event sources to return in response. This value must be greater than 0.

lesmFunctionName - The name of the Lambda function. You can specify the function name (for example, Thumbnail ) or you can specify Amazon Resource Name (ARN) of the function (for example, arn:aws:lambda:us-west-2:account-id:function:ThumbNail ). If you are using versioning, you can also provide a qualified function ARN (ARN that is qualified with function version or alias name as suffix). AWS Lambda also allows you to specify only the function name with the account ID qualifier (for example, account-id:Thumbnail ). Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 characters in length.

data ListEventSourceMappings

See: listEventSourceMappings smart constructor.

Request Lenses

lesmEventSourceARN :: Lens' ListEventSourceMappings (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Kinesis stream. (This parameter is optional.)

lesmMarker :: Lens' ListEventSourceMappings (Maybe Text)

Optional string. An opaque pagination token returned from a previous ListEventSourceMappings operation. If present, specifies to continue the list from where the returning call left off.

lesmMaxItems :: Lens' ListEventSourceMappings (Maybe Natural)

Optional integer. Specifies the maximum number of event sources to return in response. This value must be greater than 0.

lesmFunctionName :: Lens' ListEventSourceMappings (Maybe Text)

The name of the Lambda function. You can specify the function name (for example, Thumbnail ) or you can specify Amazon Resource Name (ARN) of the function (for example, arn:aws:lambda:us-west-2:account-id:function:ThumbNail ). If you are using versioning, you can also provide a qualified function ARN (ARN that is qualified with function version or alias name as suffix). AWS Lambda also allows you to specify only the function name with the account ID qualifier (for example, account-id:Thumbnail ). Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 characters in length.

Destructuring the Response

listEventSourceMappingsResponse

Arguments

:: Int
lesmrsResponseStatus

-> ListEventSourceMappingsResponse

Creates a value of ListEventSourceMappingsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lesmrsEventSourceMappings - An array of EventSourceMappingConfiguration objects.

lesmrsNextMarker - A string, present if there are more event source mappings.

lesmrsResponseStatus - -- | The response status code.

data ListEventSourceMappingsResponse

Contains a list of event sources (see EventSourceMappingConfiguration )

See: listEventSourceMappingsResponse smart constructor.

Response Lenses

lesmrsEventSourceMappings :: Lens' ListEventSourceMappingsResponse [EventSourceMappingConfiguration]

An array of EventSourceMappingConfiguration objects.

lesmrsNextMarker :: Lens' ListEventSourceMappingsResponse (Maybe Text)

A string, present if there are more event source mappings.

lesmrsResponseStatus :: Lens' ListEventSourceMappingsResponse Int

- | The response status code.