Network.AWS.Lambda.ListFunctions

Description

Returns a list of your Lambda functions. For each function, the response includes the function configuration information. You must use GetFunction to retrieve the code for your function.

This operation requires permission for the lambda:ListFunctions action.

If you are using the versioning feature, you can list all of your functions or only > LATEST versions. For information about the versioning feature, see AWS Lambda Function Versioning and Aliases .

This operation returns paginated results.

Creating a Request

listFunctions :: ListFunctions

Creates a value of ListFunctions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lfMasterRegion - Optional string. If not specified, will return only regular function versions (i.e., non-replicated versions). Valid values are: The region from which the functions are replicated. For example, if you specify us-east-1 , only functions replicated from that region will be returned. ALL _ Will return all functions from any region. If specified, you also must specify a valid FunctionVersion parameter.

lfMarker - Optional string. An opaque pagination token returned from a previous ListFunctions operation. If present, indicates where to continue the listing.

lfMaxItems - Optional integer. Specifies the maximum number of AWS Lambda functions to return in response. This parameter value must be greater than 0.

lfFunctionVersion - Optional string. If not specified, only the unqualified functions ARNs (Amazon Resource Names) will be returned. Valid value: ALL _ Will return all versions, including > LATEST which will have fully qualified ARNs (Amazon Resource Names).

data ListFunctions

See: listFunctions smart constructor.

Request Lenses

lfMasterRegion :: Lens' ListFunctions (Maybe Text)

Optional string. If not specified, will return only regular function versions (i.e., non-replicated versions). Valid values are: The region from which the functions are replicated. For example, if you specify us-east-1 , only functions replicated from that region will be returned. ALL _ Will return all functions from any region. If specified, you also must specify a valid FunctionVersion parameter.

lfMarker :: Lens' ListFunctions (Maybe Text)

Optional string. An opaque pagination token returned from a previous ListFunctions operation. If present, indicates where to continue the listing.

lfMaxItems :: Lens' ListFunctions (Maybe Natural)

Optional integer. Specifies the maximum number of AWS Lambda functions to return in response. This parameter value must be greater than 0.

lfFunctionVersion :: Lens' ListFunctions (Maybe FunctionVersion)

Optional string. If not specified, only the unqualified functions ARNs (Amazon Resource Names) will be returned. Valid value: ALL _ Will return all versions, including > LATEST which will have fully qualified ARNs (Amazon Resource Names).

Destructuring the Response

listFunctionsResponse

Arguments

:: Int
lfrsResponseStatus

-> ListFunctionsResponse

Creates a value of ListFunctionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lfrsNextMarker - A string, present if there are more functions.

lfrsFunctions - A list of Lambda functions.

lfrsResponseStatus - -- | The response status code.

data ListFunctionsResponse

Contains a list of AWS Lambda function configurations (see FunctionConfiguration .

See: listFunctionsResponse smart constructor.

Response Lenses

lfrsNextMarker :: Lens' ListFunctionsResponse (Maybe Text)

A string, present if there are more functions.

lfrsFunctions :: Lens' ListFunctionsResponse [FunctionConfiguration]

A list of Lambda functions.

lfrsResponseStatus :: Lens' ListFunctionsResponse Int

- | The response status code.