Network.AWS.Lambda.ListFunctions

Description

Returns a list of your Lambda functions. For each function, the response includes the function configuration information. You must use GetFunction to retrieve the code for your function.

This operation requires permission for the 'lambda:ListFunctions' action.

http://docs.aws.amazon.com/lambda/latest/dg/API_ListFunctions.html

Request

data ListFunctions

Request constructor

listFunctions :: ListFunctions

ListFunctions constructor.

The fields accessible through corresponding lenses are:

lfMarker :: Maybe Text

lfMaxItems :: Maybe Natural

Request lenses

lfMarker :: Lens' ListFunctions (Maybe Text)

Optional string. An opaque pagination token returned from a previous ListFunctions operation. If present, indicates where to continue the listing.

lfMaxItems :: Lens' ListFunctions (Maybe Natural)

Optional integer. Specifies the maximum number of AWS Lambda functions to return in response. This parameter value must be greater than 0.

Response

data ListFunctionsResponse

Response constructor

listFunctionsResponse :: ListFunctionsResponse

ListFunctionsResponse constructor.

The fields accessible through corresponding lenses are:

lfrFunctions :: [FunctionConfiguration]

lfrNextMarker :: Maybe Text

Response lenses

lfrFunctions :: Lens' ListFunctionsResponse [FunctionConfiguration]

A list of Lambda functions.

lfrNextMarker :: Lens' ListFunctionsResponse (Maybe Text)

A string, present if there are more functions.