Network.AWS.Lambda.RemovePermission

Description

You can remove individual permissions from an resource policy associated with a Lambda function by providing a statement ID that you provided when you added the permission.

If you are using versioning, the permissions you remove are specific to the Lambda function version or alias you specify in the AddPermission request via the Qualifier parameter. For more information about versioning, see AWS Lambda Function Versioning and Aliases .

Note that removal of a permission will cause an active event source to lose permission to the function.

You need permission for the lambda:RemovePermission action.

Creating a Request

removePermission

Arguments

:: Text
rpFunctionName

-> Text
rpStatementId

-> RemovePermission

Creates a value of RemovePermission with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rpQualifier - You can specify this optional parameter to remove permission associated with a specific function version or function alias. If you don't specify this parameter, the API removes permission associated with the unqualified function ARN.

rpFunctionName - Lambda function whose resource policy you want to remove a permission from. You can specify a function name (for example, Thumbnail ) or you can specify Amazon Resource Name (ARN) of the function (for example, arn:aws:lambda:us-west-2:account-id:function:ThumbNail ). AWS Lambda also allows you to specify a partial ARN (for example, account-id:Thumbnail ). Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 characters in length.

rpStatementId - Statement ID of the permission to remove.

data RemovePermission

See: removePermission smart constructor.

Request Lenses

rpQualifier :: Lens' RemovePermission (Maybe Text)

You can specify this optional parameter to remove permission associated with a specific function version or function alias. If you don't specify this parameter, the API removes permission associated with the unqualified function ARN.

rpFunctionName :: Lens' RemovePermission Text

Lambda function whose resource policy you want to remove a permission from. You can specify a function name (for example, Thumbnail ) or you can specify Amazon Resource Name (ARN) of the function (for example, arn:aws:lambda:us-west-2:account-id:function:ThumbNail ). AWS Lambda also allows you to specify a partial ARN (for example, account-id:Thumbnail ). Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 characters in length.

rpStatementId :: Lens' RemovePermission Text

Statement ID of the permission to remove.

Destructuring the Response

removePermissionResponse :: RemovePermissionResponse

Creates a value of RemovePermissionResponse with the minimum fields required to make a request.

data RemovePermissionResponse

See: removePermissionResponse smart constructor.