Network.AWS.Lambda.Types

Service

data Lambda

Version 2015-03-31 of the Amazon Lambda service.

Error

data JSONError :: *

Runtime

data Runtime

Constructors

Jvm
jvm

Nodejs
nodejs

Python
python

EventSourcePosition

data EventSourcePosition

Constructors

Latest
LATEST

TrimHorizon
TRIM_HORIZON

InvocationType

data InvocationType

Constructors

DryRun
DryRun

Event
Event

RequestResponse
RequestResponse

LogType

data LogType

Constructors

None
None

Tail'
Tail

FunctionCode

data FunctionCode

functionCode :: FunctionCode

FunctionCode constructor.

The fields accessible through corresponding lenses are:

fcZipFile :: Maybe Base64

fcZipFile :: Lens' FunctionCode (Maybe Base64)

A base64-encoded .zip file containing your packaged source code. For more information about creating a .zip file, go to Execution Permissions in the AWS Lambda Developer Guide.

FunctionCodeLocation

data FunctionCodeLocation

functionCodeLocation :: FunctionCodeLocation

FunctionCodeLocation constructor.

The fields accessible through corresponding lenses are:

fclLocation :: Maybe Text

fclRepositoryType :: Maybe Text

fclLocation :: Lens' FunctionCodeLocation (Maybe Text)

The presigned URL you can use to download the function's .zip file that you previously uploaded. The URL is valid for up to 10 minutes.

fclRepositoryType :: Lens' FunctionCodeLocation (Maybe Text)

The repository from which you can download the function.

FunctionConfiguration

data FunctionConfiguration

functionConfiguration :: FunctionConfiguration

FunctionConfiguration constructor.

The fields accessible through corresponding lenses are:

fcCodeSize :: Maybe Integer

fcDescription :: Maybe Text

fcFunctionArn :: Maybe Text

fcFunctionName :: Maybe Text

fcHandler :: Maybe Text

fcLastModified :: Maybe Text

fcMemorySize :: Maybe Natural

fcRole :: Maybe Text

fcRuntime :: Maybe Runtime

fcTimeout :: Maybe Natural

fcCodeSize :: Lens' FunctionConfiguration (Maybe Integer)

The size, in bytes, of the function .zip file you uploaded.

fcDescription :: Lens' FunctionConfiguration (Maybe Text)

The user-provided description.

fcFunctionArn :: Lens' FunctionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) assigned to the function.

fcFunctionName :: Lens' FunctionConfiguration (Maybe Text)

The name of the function.

fcHandler :: Lens' FunctionConfiguration (Maybe Text)

The function Lambda calls to begin executing your function.

fcLastModified :: Lens' FunctionConfiguration (Maybe Text)

The timestamp of the last time you updated the function.

fcMemorySize :: Lens' FunctionConfiguration (Maybe Natural)

The memory size, in MB, you configured for the function. Must be a multiple of 64 MB.

fcRole :: Lens' FunctionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources.

fcRuntime :: Lens' FunctionConfiguration (Maybe Runtime)

The runtime environment for the Lambda function.

fcTimeout :: Lens' FunctionConfiguration (Maybe Natural)

The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

EventSourceMappingConfiguration

data EventSourceMappingConfiguration

eventSourceMappingConfiguration :: EventSourceMappingConfiguration

EventSourceMappingConfiguration constructor.

The fields accessible through corresponding lenses are:

esmcBatchSize :: Maybe Natural

esmcEventSourceArn :: Maybe Text

esmcFunctionArn :: Maybe Text

esmcLastModified :: Maybe UTCTime

esmcLastProcessingResult :: Maybe Text

esmcState :: Maybe Text

esmcStateTransitionReason :: Maybe Text

esmcUUID :: Maybe Text

esmcBatchSize :: Lens' EventSourceMappingConfiguration (Maybe Natural)

The largest number of records that AWS Lambda will retrieve from your event source at the time of invoking your function. Your function receives an event with all the retrieved records.

esmcEventSourceArn :: Lens' EventSourceMappingConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Kinesis stream that is the source of events.

esmcFunctionArn :: Lens' EventSourceMappingConfiguration (Maybe Text)

The Lambda function to invoke when AWS Lambda detects an event on the stream.

esmcLastModified :: Lens' EventSourceMappingConfiguration (Maybe UTCTime)

The UTC time string indicating the last time the event mapping was updated.

esmcLastProcessingResult :: Lens' EventSourceMappingConfiguration (Maybe Text)

The result of the last AWS Lambda invocation of your Lambda function.

esmcState :: Lens' EventSourceMappingConfiguration (Maybe Text)

The state of the event source mapping. It can be Creating, Enabled, Disabled, Enabling, Disabling, Updating, or Deleting.

esmcStateTransitionReason :: Lens' EventSourceMappingConfiguration (Maybe Text)

The reason the event source mapping is in its current state. It is either user-requested or an AWS Lambda-initiated state transition.

esmcUUID :: Lens' EventSourceMappingConfiguration (Maybe Text)

The AWS Lambda assigned opaque identifier for the mapping.