Network.AWS.Lambda.UpdateFunctionConfiguration

Description

Updates the configuration parameters for the specified Lambda function by using the values provided in the request. You provide only the parameters you want to change. This operation must only be used on an existing Lambda function and cannot be used to update the function's code.

If you are using the versioning feature, note this API will always update the $LATEST version of your Lambda function. For information about the versioning feature, see AWS Lambda Function Versioning and Aliases .

This operation requires permission for the lambda:UpdateFunctionConfiguration action.

Creating a Request

updateFunctionConfiguration

Arguments

:: Text
ufcFunctionName

-> UpdateFunctionConfiguration

Creates a value of UpdateFunctionConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ufcMemorySize - The amount of memory, in MB, your Lambda function is given. AWS Lambda uses this memory size to infer the amount of CPU allocated to your function. Your function use-case determines your CPU and memory requirements. For example, a database operation might need less memory compared to an image processing function. The default value is 128 MB. The value must be a multiple of 64 MB.

ufcRuntime - The runtime environment for the Lambda function. To use the Python runtime v3.6, set the value to "python3.6". To use the Python runtime v2.7, set the value to "python2.7". To use the Node.js runtime v6.10, set the value to "nodejs6.10". To use the Node.js runtime v4.3, set the value to "nodejs4.3". To use the Python runtime v3.6, set the value to "python3.6".

ufcKMSKeyARN - The Amazon Resource Name (ARN) of the KMS key used to encrypt your function's environment variables. If you elect to use the AWS Lambda default service key, pass in an empty string ("") for this parameter.

ufcEnvironment - The parent object that contains your environment's configuration settings.

ufcDeadLetterConfig - The parent object that contains the target ARN (Amazon Resource Name) of an Amazon SQS queue or Amazon SNS topic.

ufcRole - The Amazon Resource Name (ARN) of the IAM role that Lambda will assume when it executes your function.

ufcVPCConfig - Undocumented member.

ufcHandler - The function that Lambda calls to begin executing your function. For Node.js, it is the module-name.export value in your function.

ufcTimeout - The function execution time at which AWS Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

ufcTracingConfig - The parent object that contains your function's tracing settings.

ufcDescription - A short user-defined function description. AWS Lambda does not use this value. Assign a meaningful description as you see fit.

ufcFunctionName - The name of the Lambda function. You can specify a function name (for example, Thumbnail ) or you can specify Amazon Resource Name (ARN) of the function (for example, arn:aws:lambda:us-west-2:account-id:function:ThumbNail ). AWS Lambda also allows you to specify a partial ARN (for example, account-id:Thumbnail ). Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 character in length.

data UpdateFunctionConfiguration

See: updateFunctionConfiguration smart constructor.

Request Lenses

ufcMemorySize :: Lens' UpdateFunctionConfiguration (Maybe Natural)

The amount of memory, in MB, your Lambda function is given. AWS Lambda uses this memory size to infer the amount of CPU allocated to your function. Your function use-case determines your CPU and memory requirements. For example, a database operation might need less memory compared to an image processing function. The default value is 128 MB. The value must be a multiple of 64 MB.

ufcRuntime :: Lens' UpdateFunctionConfiguration (Maybe Runtime)

The runtime environment for the Lambda function. To use the Python runtime v3.6, set the value to "python3.6". To use the Python runtime v2.7, set the value to "python2.7". To use the Node.js runtime v6.10, set the value to "nodejs6.10". To use the Node.js runtime v4.3, set the value to "nodejs4.3". To use the Python runtime v3.6, set the value to "python3.6".

ufcKMSKeyARN :: Lens' UpdateFunctionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the KMS key used to encrypt your function's environment variables. If you elect to use the AWS Lambda default service key, pass in an empty string ("") for this parameter.

ufcEnvironment :: Lens' UpdateFunctionConfiguration (Maybe Environment)

The parent object that contains your environment's configuration settings.

ufcDeadLetterConfig :: Lens' UpdateFunctionConfiguration (Maybe DeadLetterConfig)

The parent object that contains the target ARN (Amazon Resource Name) of an Amazon SQS queue or Amazon SNS topic.

ufcRole :: Lens' UpdateFunctionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that Lambda will assume when it executes your function.

ufcVPCConfig :: Lens' UpdateFunctionConfiguration (Maybe VPCConfig)

Undocumented member.

ufcHandler :: Lens' UpdateFunctionConfiguration (Maybe Text)

The function that Lambda calls to begin executing your function. For Node.js, it is the module-name.export value in your function.

ufcTimeout :: Lens' UpdateFunctionConfiguration (Maybe Natural)

The function execution time at which AWS Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

ufcTracingConfig :: Lens' UpdateFunctionConfiguration (Maybe TracingConfig)

The parent object that contains your function's tracing settings.

ufcDescription :: Lens' UpdateFunctionConfiguration (Maybe Text)

A short user-defined function description. AWS Lambda does not use this value. Assign a meaningful description as you see fit.

ufcFunctionName :: Lens' UpdateFunctionConfiguration Text

The name of the Lambda function. You can specify a function name (for example, Thumbnail ) or you can specify Amazon Resource Name (ARN) of the function (for example, arn:aws:lambda:us-west-2:account-id:function:ThumbNail ). AWS Lambda also allows you to specify a partial ARN (for example, account-id:Thumbnail ). Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 character in length.

Destructuring the Response

functionConfiguration :: FunctionConfiguration

Creates a value of FunctionConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fcMemorySize - The memory size, in MB, you configured for the function. Must be a multiple of 64 MB.

fcRuntime - The runtime environment for the Lambda function.

fcFunctionARN - The Amazon Resource Name (ARN) assigned to the function.

fcKMSKeyARN - The Amazon Resource Name (ARN) of the KMS key used to encrypt your function's environment variables. If empty, it means you are using the AWS Lambda default service key.

fcEnvironment - The parent object that contains your environment's configuration settings.

fcDeadLetterConfig - The parent object that contains the target ARN (Amazon Resource Name) of an Amazon SQS queue or Amazon SNS topic.

fcRole - The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources.

fcVPCConfig - VPC configuration associated with your Lambda function.

fcVersion - The version of the Lambda function.

fcFunctionName - The name of the function. Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 characters in length.

fcCodeSize - The size, in bytes, of the function .zip file you uploaded.

fcHandler - The function Lambda calls to begin executing your function.

fcTimeout - The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

fcLastModified - The time stamp of the last time you updated the function. The time stamp is conveyed as a string complying with ISO-8601 in this way YYYY-MM-DDThh:mm:ssTZD (e.g., 1997-07-16T19:20:30+01:00). For more information, see Date and Time Formats .

fcCodeSha256 - It is the SHA256 hash of your function deployment package.

fcTracingConfig - The parent object that contains your function's tracing settings.

fcDescription - The user-provided description.

fcMasterARN - Returns the ARN (Amazon Resource Name) of the master function.

data FunctionConfiguration

A complex type that describes function metadata.

See: functionConfiguration smart constructor.

Response Lenses

fcMemorySize :: Lens' FunctionConfiguration (Maybe Natural)

The memory size, in MB, you configured for the function. Must be a multiple of 64 MB.

fcRuntime :: Lens' FunctionConfiguration (Maybe Runtime)

The runtime environment for the Lambda function.

fcFunctionARN :: Lens' FunctionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) assigned to the function.

fcKMSKeyARN :: Lens' FunctionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the KMS key used to encrypt your function's environment variables. If empty, it means you are using the AWS Lambda default service key.

fcEnvironment :: Lens' FunctionConfiguration (Maybe EnvironmentResponse)

The parent object that contains your environment's configuration settings.

fcDeadLetterConfig :: Lens' FunctionConfiguration (Maybe DeadLetterConfig)

The parent object that contains the target ARN (Amazon Resource Name) of an Amazon SQS queue or Amazon SNS topic.

fcRole :: Lens' FunctionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources.

fcVPCConfig :: Lens' FunctionConfiguration (Maybe VPCConfigResponse)

VPC configuration associated with your Lambda function.

fcVersion :: Lens' FunctionConfiguration (Maybe Text)

The version of the Lambda function.

fcFunctionName :: Lens' FunctionConfiguration (Maybe Text)

The name of the function. Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 characters in length.

fcCodeSize :: Lens' FunctionConfiguration (Maybe Integer)

The size, in bytes, of the function .zip file you uploaded.

fcHandler :: Lens' FunctionConfiguration (Maybe Text)

The function Lambda calls to begin executing your function.

fcTimeout :: Lens' FunctionConfiguration (Maybe Natural)

The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

fcLastModified :: Lens' FunctionConfiguration (Maybe Text)

The time stamp of the last time you updated the function. The time stamp is conveyed as a string complying with ISO-8601 in this way YYYY-MM-DDThh:mm:ssTZD (e.g., 1997-07-16T19:20:30+01:00). For more information, see Date and Time Formats .

fcCodeSha256 :: Lens' FunctionConfiguration (Maybe Text)

It is the SHA256 hash of your function deployment package.

fcTracingConfig :: Lens' FunctionConfiguration (Maybe TracingConfigResponse)

The parent object that contains your function's tracing settings.

fcDescription :: Lens' FunctionConfiguration (Maybe Text)

The user-provided description.

fcMasterARN :: Lens' FunctionConfiguration (Maybe Text)

Returns the ARN (Amazon Resource Name) of the master function.