Network.AWS.Lambda.UpdateFunctionConfiguration

Description

Updates the configuration parameters for the specified Lambda function by using the values provided in the request. You provide only the parameters you want to change. This operation must only be used on an existing Lambda function and cannot be used to update the function's code.

This operation requires permission for the 'lambda:UpdateFunctionConfiguration' action.

http://docs.aws.amazon.com/lambda/latest/dg/API_UpdateFunctionConfiguration.html

Request

data UpdateFunctionConfiguration

Request constructor

updateFunctionConfiguration

Arguments

:: Text
ufcFunctionName

-> UpdateFunctionConfiguration

UpdateFunctionConfiguration constructor.

The fields accessible through corresponding lenses are:

ufcDescription :: Maybe Text

ufcFunctionName :: Text

ufcHandler :: Maybe Text

ufcMemorySize :: Maybe Natural

ufcRole :: Maybe Text

ufcTimeout :: Maybe Natural

Request lenses

ufcDescription :: Lens' UpdateFunctionConfiguration (Maybe Text)

A short user-defined function description. AWS Lambda does not use this value. Assign a meaningful description as you see fit.

ufcFunctionName :: Lens' UpdateFunctionConfiguration Text

The name of the Lambda function.

You can specify an unqualified function name (for example, Thumbnail) or you can specify Amazon Resource Name (ARN) of the function (for example, "arn:aws:lambda:us-west-2:account-id:function:ThumbNail"). AWS Lambda also allows you to specify only the account ID qualifier (for example, "account-id:Thumbnail"). Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 character in length.

ufcHandler :: Lens' UpdateFunctionConfiguration (Maybe Text)

The function that Lambda calls to begin executing your function. For Node.js, it is the module-name.export value in your function.

ufcMemorySize :: Lens' UpdateFunctionConfiguration (Maybe Natural)

The amount of memory, in MB, your Lambda function is given. AWS Lambda uses this memory size to infer the amount of CPU allocated to your function. Your function use-case determines your CPU and memory requirements. For example, a database operation might need less memory compared to an image processing function. The default value is 128 MB. The value must be a multiple of 64 MB.

ufcRole :: Lens' UpdateFunctionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that Lambda will assume when it executes your function.

ufcTimeout :: Lens' UpdateFunctionConfiguration (Maybe Natural)

The function execution time at which AWS Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

Response

data UpdateFunctionConfigurationResponse

Response constructor

updateFunctionConfigurationResponse :: UpdateFunctionConfigurationResponse

UpdateFunctionConfigurationResponse constructor.

The fields accessible through corresponding lenses are:

ufcr1CodeSize :: Maybe Integer

ufcr1Description :: Maybe Text

ufcr1FunctionArn :: Maybe Text

ufcr1FunctionName :: Maybe Text

ufcr1Handler :: Maybe Text

ufcr1LastModified :: Maybe Text

ufcr1MemorySize :: Maybe Natural

ufcr1Role :: Maybe Text

ufcr1Runtime :: Maybe Runtime

ufcr1Timeout :: Maybe Natural

Response lenses

ufcr1CodeSize :: Lens' UpdateFunctionConfigurationResponse (Maybe Integer)

The size, in bytes, of the function .zip file you uploaded.

ufcr1Description :: Lens' UpdateFunctionConfigurationResponse (Maybe Text)

The user-provided description.

ufcr1FunctionArn :: Lens' UpdateFunctionConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the function.

ufcr1FunctionName :: Lens' UpdateFunctionConfigurationResponse (Maybe Text)

The name of the function.

ufcr1Handler :: Lens' UpdateFunctionConfigurationResponse (Maybe Text)

The function Lambda calls to begin executing your function.

ufcr1LastModified :: Lens' UpdateFunctionConfigurationResponse (Maybe Text)

The timestamp of the last time you updated the function.

ufcr1MemorySize :: Lens' UpdateFunctionConfigurationResponse (Maybe Natural)

The memory size, in MB, you configured for the function. Must be a multiple of 64 MB.

ufcr1Role :: Lens' UpdateFunctionConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources.

ufcr1Runtime :: Lens' UpdateFunctionConfigurationResponse (Maybe Runtime)

The runtime environment for the Lambda function.

ufcr1Timeout :: Lens' UpdateFunctionConfigurationResponse (Maybe Natural)

The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.